% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquisition_query.R
\name{acquisition_query}
\alias{acquisition_query}
\title{Acquisition Query}
\usage{
acquisition_query(acquisition_id)
}
\arguments{
\item{acquisition_id}{integer; Internal, the unique numeric id of this acquisition.}
}
\description{
Returns \strong{information about the acquisition} corresponding to the given id.
}
