% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcquisitionModel.R
\name{AcquisitionModel}
\alias{AcquisitionModel}
\title{AcquisitionModel}
\usage{
AcquisitionModel(
  acquisid,
  acq_sample_id,
  orig_id,
  instrument = NULL,
  free_columns = NULL
)
}
\arguments{
\item{acquisid}{[integer, required] The acquisition Id.}

\item{acq_sample_id}{[integer, required] The acquisition sample Id.}

\item{orig_id}{[string, required] Original acquisition ID from initial TSV load.}

\item{instrument}{[string] Instrument used.}

\item{free_columns}{[list] Free columns from acquisition mapping in project.}
}
\description{
A list defining a AcquisitionModel
}
