% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquisitions_search.R
\name{acquisitions_search}
\alias{acquisitions_search}
\title{Acquisitions Search}
\usage{
acquisitions_search(project_id)
}
\arguments{
\item{project_id}{integer; The project id.}
}
\description{
Returns the \strong{list of all acquisitions for a given project}.
}
