% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_user.R
\name{activate_user}
\alias{activate_user}
\title{Activate User}
\usage{
activate_user(user_id, status, no_bot = NULL, activatereq)
}
\arguments{
\item{user_id}{integer; Internal, the unique numeric id of this user.}

\item{status}{string; Internal, the status name assign to this user.}

\item{no_bot}{array; not-a-robot proof}

\item{activatereq}{NULL; activation parameters : reason if the request comes from an Administrator, password if the request comes from the user as it serves as email confirmation request.}
}
\description{
Activate a new user if external validation is on., return \strong{NULL upon success.} 🔒 Depending on logged user, different authorizations apply: - An administrator or user administrator can activate a user or bypass the activation and inform the user when a modification request value/reason is provided. - An ordinary logged user cannot activate another account. If back-end configuration for self-creation check is Google reCAPTCHA, then no_bot is a pair \link{remote IP, reCAPTCHA response}.
}
