% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_worms_taxon.R
\name{add_worms_taxon}
\alias{add_worms_taxon}
\title{Add Worms Taxon}
\usage{
add_worms_taxon(AddWormsTaxonModel)
}
\arguments{
\item{AddWormsTaxonModel}{Output of \code{\link[=AddWormsTaxonModel]{AddWormsTaxonModel()}};}
}
\description{
Add worms taxon by its aphia_id
}
