% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddWormsTaxonModel.R
\name{AddWormsTaxonModel}
\alias{AddWormsTaxonModel}
\title{AddWormsTaxonModel}
\usage{
AddWormsTaxonModel(aphia_id = NULL)
}
\arguments{
\item{aphia_id}{[integer] The unique numeric aphia_id of the taxon in WoRMS.}
}
\description{
A list defining a AddWormsTaxonModel
}
