% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackupExportReq.R
\name{BackupExportReq}
\alias{BackupExportReq}
\title{Backup Export request Model}
\usage{
BackupExportReq(collection_id = NULL, project_id, out_to_ftp = NULL)
}
\arguments{
\item{collection_id}{[integer] The Collection to export if requested.}

\item{project_id}{[integer, string, required] The project(int) or projects (str, project ids list) to export.}

\item{out_to_ftp}{[boolean] Copy result file to FTP area. Original file is still available.}
}
\description{
A list defining a Backup Export request Model
}
