% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Body_create_file_user_files_create__post.R
\name{Body_create_file_user_files_create__post}
\alias{Body_create_file_user_files_create__post}
\title{Body_create_file_user_files_create__post}
\usage{
Body_create_file_user_files_create__post(source_path = NULL)
}
\arguments{
\item{source_path}{[string] The path of the file or directory to be moved.}
}
\description{
A list defining a Body_create_file_user_files_create__post
}
