% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Body_export_object_set_general_object_set_export_general_post.R
\name{Body_export_object_set_general_object_set_export_general_post}
\alias{Body_export_object_set_general_object_set_export_general_post}
\title{Body_export_object_set_general_object_set_export_general_post}
\usage{
Body_export_object_set_general_object_set_export_general_post(filters, request)
}
\arguments{
\item{filters}{[object of type \link{ProjectFilters}, required] filters}

\item{request}{[object of type \link{GeneralExportReq}, required] request}
}
\description{
A list defining a Body_export_object_set_general_object_set_export_general_post
}
