% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Body_export_object_set_object_set_export_post.R
\name{Body_export_object_set_object_set_export_post}
\alias{Body_export_object_set_object_set_export_post}
\title{Body_export_object_set_object_set_export_post}
\usage{
Body_export_object_set_object_set_export_post(filters, request)
}
\arguments{
\item{filters}{[object of type \link{ProjectFilters}, required] filters}

\item{request}{[object of type \link{ExportReq}, required] request}
}
\description{
A list defining a Body_export_object_set_object_set_export_post
}
