% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Body_move_file_user_files_mv__post.R
\name{Body_move_file_user_files_mv__post}
\alias{Body_move_file_user_files_mv__post}
\title{Body_move_file_user_files_mv__post}
\usage{
Body_move_file_user_files_mv__post(source_path = NULL, dest_path = NULL)
}
\arguments{
\item{source_path}{[string] The  path of the file or directory to be moved.}

\item{dest_path}{[string] The path of the destination file or directory.}
}
\description{
A list defining a Body_move_file_user_files_mv__post
}
