% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Body_predict_object_set_object_set_predict_post.R
\name{Body_predict_object_set_object_set_predict_post}
\alias{Body_predict_object_set_object_set_predict_post}
\title{Body_predict_object_set_object_set_predict_post}
\usage{
Body_predict_object_set_object_set_predict_post(filters, request)
}
\arguments{
\item{filters}{[object of type \link{ProjectFilters}, required] filters}

\item{request}{[object of type \link{PredictionReq}, required] request}
}
\description{
A list defining a Body_predict_object_set_object_set_predict_post
}
