% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Body_put_my_file_user_files__post.R
\name{Body_put_my_file_user_files__post}
\alias{Body_put_my_file_user_files__post}
\title{Body_put_my_file_user_files__post}
\usage{
Body_put_my_file_user_files__post(file, path = NULL)
}
\arguments{
\item{file}{[string, required] File}

\item{path}{[string] The destination path of the file.}
}
\description{
A list defining a Body_put_my_file_user_files__post
}
