% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Body_put_user_file_my_files_post.R
\name{Body_put_user_file_my_files_post}
\alias{Body_put_user_file_my_files_post}
\title{Body_put_user_file_my_files_post}
\usage{
Body_put_user_file_my_files_post(file, path = NULL, tag = NULL)
}
\arguments{
\item{file}{[string, required] File}

\item{path}{[string] The client-side full path of the file.}

\item{tag}{[string] If a tag is provided, then all files with the same tag are grouped (in a sub-directory). Otherwise, a temp directory with only this file will be created.}
}
\description{
A list defining a Body_put_user_file_my_files_post
}
