% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Body_remove_file_user_files_rm__post.R
\name{Body_remove_file_user_files_rm__post}
\alias{Body_remove_file_user_files_rm__post}
\title{Body_remove_file_user_files_rm__post}
\usage{
Body_remove_file_user_files_rm__post(source_path = NULL)
}
\arguments{
\item{source_path}{[string] The path of the file  or directory to be removed. * for all files and directories}
}
\description{
A list defining a Body_remove_file_user_files_rm__post
}
