% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkUpdateReq.R
\name{BulkUpdateReq}
\alias{BulkUpdateReq}
\title{Update request Model}
\usage{
BulkUpdateReq(target_ids, updates)
}
\arguments{
\item{target_ids}{[vector of integer, required] The IDs of the target entities.}

\item{updates}{[vector of , required] The list of updates, to do on all impacted entities.        {            ucol : A column name, pseudo-columns AKA free ones, are OK.            uval : The new value to set, always as a string        }}
}
\description{
A list defining a Update request Model
}
