% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_auto_mult_object_set.R
\name{classify_auto_mult_object_set}
\alias{classify_auto_mult_object_set}
\title{Classify Auto Mult Object Set}
\usage{
classify_auto_mult_object_set(ClassifyAutoReqMult)
}
\arguments{
\item{ClassifyAutoReqMult}{Output of \code{\link[=ClassifyAutoReqMult]{ClassifyAutoReqMult()}};}
}
\description{
\strong{Set automatic classification} of a set of objects. \strong{Returns the number of updated entities.}
}
