% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_auto_object_set.R
\name{classify_auto_object_set}
\alias{classify_auto_object_set}
\title{Classify Auto Object Set}
\usage{
classify_auto_object_set(ClassifyAutoReq)
}
\arguments{
\item{ClassifyAutoReq}{Output of ClassifyAutoReq();}
}
\description{
\strong{Set automatic classification} of a set of objects. \strong{Returns the number of updated entities.}
}
