% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_object_set.R
\name{classify_object_set}
\alias{classify_object_set}
\title{Classify Object Set}
\usage{
classify_object_set(
  target_ids = NULL,
  classifications = NULL,
  wanted_qualification = NULL
)
}
\arguments{
\item{target_ids}{array; The IDs of the target objects.}

\item{classifications}{array; The wanted new classifications, i.e. taxon ID, one for each object. Use -1 to keep present one.}

\item{wanted_qualification}{string; The wanted qualifications for all objects. 'V' or 'P'.}
}
\description{
\strong{Change classification and/or qualification for a set of objects.} \strong{Returns the number of updated entities.} 🔒 Current user needs at \emph{least Annotate} right on all projects of specified objects.
}
