% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyAutoReq.R
\name{ClassifyAutoReq}
\alias{ClassifyAutoReq}
\title{ClassifyAutoReq}
\usage{
ClassifyAutoReq(target_ids, classifications, scores, keep_log)
}
\arguments{
\item{target_ids}{[vector of integer, required] The IDs of the target objects.}

\item{classifications}{[vector of integer, required] The wanted new classifications, i.e. taxon ID, one for each object.}

\item{scores}{[vector of number, required] The classification score is generally between 0 and 1. It indicates the probability that the taxon prediction of this object is correct.}

\item{keep_log}{[boolean, required] Set if former automatic classification history is needed.}
}
\description{
A list defining a ClassifyAutoReq
}
