% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyAutoReqMult.R
\name{ClassifyAutoReqMult}
\alias{ClassifyAutoReqMult}
\title{Classify auto request Model}
\usage{
ClassifyAutoReqMult(target_ids, classifications, scores, keep_log)
}
\arguments{
\item{target_ids}{[vector of integer, required] The IDs of the target objects.}

\item{classifications}{[vector of array, required] The wanted new classifications, i.e. taxon ID, one list for each object.}

\item{scores}{[vector of array, required] The classification scores, between 0 and 1. Each indicates the probability that the taxon prediction of this object for this category is correct.}

\item{keep_log}{[boolean, required] Set if former automatic classification history is needed. Deprecated, always True.}
}
\description{
A list defining a Classify auto request Model
}
