% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyReq.R
\name{ClassifyReq}
\alias{ClassifyReq}
\title{ClassifyReq}
\usage{
ClassifyReq(target_ids, classifications, wanted_qualification)
}
\arguments{
\item{target_ids}{[vector of integer, required] The IDs of the target objects.}

\item{classifications}{[vector of integer, required] The wanted new classifications, i.e. taxon ID, one for each object. Use -1 to keep present one.}

\item{wanted_qualification}{[string, required] The wanted qualifications for all objects. 'V' or 'P'.}
}
\description{
A list defining a ClassifyReq
}
