% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_aggregated_projects_properties.R
\name{collection_aggregated_projects_properties}
\alias{collection_aggregated_projects_properties}
\title{Collection Aggregated Projects Properties}
\usage{
collection_aggregated_projects_properties(project_ids)
}
\arguments{
\item{project_ids}{string; String containing the list of one or more project id separated by non-num char. .}
}
\description{
**returns projectset calculated selected fields values projects and list of rejected projects id. Note: 'manage' right is required on all underlying projects.
}
