% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_by_short_title.R
\name{collection_by_short_title}
\alias{collection_by_short_title}
\title{Collection By Short Title}
\usage{
collection_by_short_title(q)
}
\arguments{
\item{q}{string; Search by \strong{exact} short title.}
}
\description{
Return the \strong{single collection with this short title}. \emph{For published datasets.} ⚠️ DO NOT MODIFY BEHAVIOR ⚠️
}
