% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_by_title.R
\name{collection_by_title}
\alias{collection_by_title}
\title{Collection By Title}
\usage{
collection_by_title(q)
}
\arguments{
\item{q}{string; Search by \strong{exact} title.}
}
\description{
Return the \strong{single collection with this title}. \emph{For published datasets.} ⚠️ DO NOT MODIFY BEHAVIOR ⚠️
}
