% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectionAggregatedRsp.R
\name{CollectionAggregatedRsp}
\alias{CollectionAggregatedRsp}
\title{CollectionAggregatedRsp}
\usage{
CollectionAggregatedRsp(
  can_be_administered = NULL,
  instrument = NULL,
  access = NULL,
  initclassiflist,
  classiffieldlist,
  cnn_network_id,
  status,
  creator_users = NULL,
  privileges = NULL,
  freecols = NULL,
  excluded = NULL
)
}
\arguments{
\item{can_be_administered}{[boolean] Whether the current user is manager of all collection projects}

\item{instrument}{[string] The collection instrument from projects.}

\item{access}{[object of type \link{AccessLevelEnum}] The restricted access for collection projects.PUBLIC: "1", OPEN: "2", PRIVATE: "0"}

\item{initclassiflist}{[string, required]  Aggregated categories from the collection projects.}

\item{classiffieldlist}{[string, required]  Aggregated sorting and displaying fields from the collection projects.}

\item{cnn_network_id}{[string, required]  Common deep features for the collection projects. Can be None (???)}

\item{status}{[string, required]  the restricted collection status calculated from projects.}

\item{creator_users}{[vector of ] Annotators extracted from history.}

\item{privileges}{[list] Aggregated user privileges of projects with user minimal right on projects}

\item{freecols}{[list] Common free cols of projects.}

\item{excluded}{[list] Excluded projects for common values}
}
\description{
A list defining a CollectionAggregatedRsp
}
