% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectionModel.R
\name{CollectionModel}
\alias{CollectionModel}
\title{Collection Model}
\usage{
CollectionModel(
  id,
  external_id,
  external_id_system,
  title,
  short_title = NULL,
  citation = NULL,
  license = NULL,
  abstract = NULL,
  description = NULL,
  project_ids,
  provider_user = NULL,
  contact_user = NULL,
  creator_users = NULL,
  creator_organisations = NULL,
  associate_users = NULL,
  associate_organisations = NULL
)
}
\arguments{
\item{id}{[integer, required] The collection Id.}

\item{external_id}{[string, required] The external Id.}

\item{external_id_system}{[string, required] The external Id system.}

\item{title}{[string, required] The collection title.}

\item{short_title}{[string] The collection short title.}

\item{citation}{[string] The collection citation.}

\item{license}{[string] The collection license.}

\item{abstract}{[string] The collection abstract.}

\item{description}{[string] The collection description.}

\item{project_ids}{[vector of integer, required] The list of composing project IDs.}

\item{provider_user}{[object of type \link{MinUserModel}] Is the person who         is responsible for the content of this metadata record. Writer of the title and abstract.}

\item{contact_user}{[object of type \link{MinUserModel}] Is the person who         should be contacted in cases of questions regarding the content of the dataset or any data restrictions.         This is also the person who is most likely to stay involved in the dataset the longest.}

\item{creator_users}{[vector of ] All people who         are responsible for the creation of the collection. Data creators should receive credit         for their work and should therefore be included in the citation.}

\item{creator_organisations}{[vector of ] All         organisations who are responsible for the creation of the collection. Data creators should         receive credit for their work and should therefore be included in the citation.}

\item{associate_users}{[vector of ] Other person(s)         associated with the collection.}

\item{associate_organisations}{[vector of ] Other         organisation(s) associated with the collection.}
}
\description{
A list defining a Collection Model
}
