% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectionReq.R
\name{CollectionReq}
\alias{CollectionReq}
\title{Update full or partial Collection Model}
\usage{
CollectionReq(
  external_id = NULL,
  title = NULL,
  short_title = NULL,
  provider_user = NULL,
  contact_user = NULL,
  project_ids = NULL,
  license = NULL,
  citation = NULL,
  abstract = NULL,
  description = NULL,
  creator_users = NULL,
  associate_users = NULL,
  creator_organisations = NULL,
  associate_organisations = NULL
)
}
\arguments{
\item{external_id}{[string] The external Id.}

\item{title}{[string] The collection title.}

\item{short_title}{[string] The collection short title.}

\item{provider_user}{[integer] Id of the person who         is responsible for the content of this metadata record. Writer of the title and abstract.}

\item{contact_user}{[integer] Id of the person who         should be contacted in cases of questions regarding the content of the dataset or any data restrictions.         This is also the person who is most likely to stay involved in the dataset the longest.}

\item{project_ids}{[vector of integer] The list of composing project IDs.}

\item{license}{[object of type \link{LicenseEnum}] The collection license.}

\item{citation}{[string] The collection citation.}

\item{abstract}{[string] The collection abstract.}

\item{description}{[string] The collection description.}

\item{creator_users}{[vector of ] List of users id's or dict with name, organization for external persons.}

\item{associate_users}{[vector of ] List of users id's or dict with name, organization for external persons.}

\item{creator_organisations}{[vector of ] All         organisations who are responsible for the creation of the collection. Data creators should         receive credit for their work and should therefore be included in the citation.}

\item{associate_organisations}{[vector of ] Other         organisation(s) Ids or names associated with the collection.}
}
\description{
A list defining a Update full or partial Collection Model
}
