% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColUpdate.R
\name{ColUpdate}
\alias{ColUpdate}
\title{Update request Model}
\usage{
ColUpdate(ucol, uval)
}
\arguments{
\item{ucol}{[string, required] Ucol}

\item{uval}{[string, required] Uval}
}
\description{
A list defining a Update request Model
}
