% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\name{Constants}
\alias{Constants}
\title{Constants}
\usage{
Constants(license_texts = NULL, app_manager = NULL, countries = NULL)
}
\arguments{
\item{license_texts}{[list] The supported licenses and help text/links.}

\item{app_manager}{[vector of string] The application manager identity (name, mail), from config file.}

\item{countries}{[vector of string] List of known countries names.}
}
\description{
A list defining a Constants
}
