% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_collection.R
\name{create_collection}
\alias{create_collection}
\title{Create Collection}
\usage{
create_collection(CreateCollectionReq)
}
\arguments{
\item{CreateCollectionReq}{Output of CreateCollectionReq();}
}
\description{
\strong{Create a collection} with at least one project inside. Returns the created collection Id. Note: 'manage' right is required on all underlying projects.
}
