% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_guest.R
\name{create_guest}
\alias{create_guest}
\title{Create Guest}
\usage{
create_guest(GuestModel)
}
\arguments{
\item{GuestModel}{Output of \code{\link[=GuestModel]{GuestModel()}};}
}
\description{
\strong{Create a new guest}, return \strong{NULL upon success.} 🔒 Depending on logged user, different authorizations apply: - An administrator or user administrator or logged project manager can create a guest. - An ordinary logged user cannot create another guest. If back-end configuration for self-creation check is Google reCAPTCHA, then no_bot is a pair \link{remote IP, reCAPTCHA response}.
}
