% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_my_file.R
\name{create_my_file}
\alias{create_my_file}
\title{Create File}
\usage{
create_my_file()
}
\description{
\strong{Create a new file or directory in the current user files directory.} The returned text will contain a server-side path which is usable for some file-related operations.
}
