% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_organization.R
\name{create_organization}
\alias{create_organization}
\title{Create Organization}
\usage{
create_organization(OrganizationModel)
}
\arguments{
\item{OrganizationModel}{Output of \code{\link[=OrganizationModel]{OrganizationModel()}};}
}
\description{
\strong{Create a new organization}, return \strong{NULL upon success.} 🔒 Depending on logged user, different authorizations apply: - An administrator or user administrator or logged project manager can create an organization. - An ordinary logged user cannot create another organization this way.
}
