% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create Project}
\usage{
create_project(CreateProjectReq)
}
\arguments{
\item{CreateProjectReq}{Output of CreateProjectReq();}
}
\description{
\strong{Create an empty project with only a title,} and \strong{return the numeric id of this newly created project}. The project will be managed by current user. 🔒 The user has to be \emph{app administrator} or \emph{project creator}.
}
