% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_user.R
\name{create_user}
\alias{create_user}
\title{Create User}
\usage{
create_user(
  no_bot = NULL,
  id = NULL,
  email = NULL,
  password = NULL,
  name = NULL,
  organisation = NULL,
  active = NULL,
  country = NULL,
  usercreationdate = NULL,
  usercreationreason = NULL,
  can_do = NULL,
  last_used_projects = NULL
)
}
\arguments{
\item{no_bot}{array; not-a-robot proof}

\item{id}{integer; The unique numeric id of this user.}

\item{email}{string; User's email address, as text, used during registration.}

\item{password}{string; Encrypted (or not) password.}

\item{name}{string; User's full name, as text.}

\item{organisation}{string; User's organisation name, as text.}

\item{active}{boolean; Whether the user is still active.}

\item{country}{string; The country name, as text (but chosen in a consistent list).}

\item{usercreationdate}{string; The date of creation of the user, as text formatted according to the ISO 8601 standard.}

\item{usercreationreason}{string; Paragraph describing the usage of EcoTaxa made by the user.}

\item{can_do}{array; List of User's allowed actions : 1 create a project, 2 administrate the app, 3 administrate users, 4 create taxon.}

\item{last_used_projects}{array; List of User's last used projects.}
}
\description{
\strong{Create a new user}, return \strong{NULL upon success.} 🔒 Depending on logged user, different authorizations apply: - An administrator or user administrator can create a user. - An unlogged user can self-create an account. But must eventually provide a no-robot proof. - An ordinary logged user cannot create another account. If back-end configuration for self-creation check is Google reCAPTCHA, then no_bot is a pair \link{remote IP, reCAPTCHA response}.
}
