% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_user_file.R
\name{create_user_file}
\alias{create_user_file}
\title{Create User File}
\usage{
create_user_file()
}
\description{
\strong{Create a new file or directory in the current user files directory.} The returned text will contain a server-side path which is usable for some file-related operations.
}
