% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCollectionReq.R
\name{CreateCollectionReq}
\alias{CreateCollectionReq}
\title{Create collection request Model}
\usage{
CreateCollectionReq(title, project_ids)
}
\arguments{
\item{title}{[string, required] The collection title.}

\item{project_ids}{[vector of integer, required] The list of composing project IDs.}
}
\description{
A list defining a Create collection request Model
}
