% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_direct_query.R
\name{db_direct_query}
\alias{db_direct_query}
\title{Direct Db Query}
\usage{
db_direct_query(q)
}
\arguments{
\item{q}{string; The SQL to execute.}
}
\description{
For making selects on the DB. 🔒 Admin only.
}
