% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectoryEntryModel.R
\name{DirectoryEntryModel}
\alias{DirectoryEntryModel}
\title{DirectoryEntryModel}
\usage{
DirectoryEntryModel(name, type, size, mtime)
}
\arguments{
\item{name}{[string, required] atomic entry name.}

\item{type}{[string, required] entry type, 'D' for directory, 'F' for file.}

\item{size}{[integer, required] Entry size, for zips.}

\item{mtime}{[string, required] Modification time, in ISO format.}
}
\description{
A list defining a DirectoryEntryModel
}
