% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectoryModel.R
\name{DirectoryModel}
\alias{DirectoryModel}
\title{DirectoryModel}
\usage{
DirectoryModel(path, entries)
}
\arguments{
\item{path}{[string, required] A /-separated path from root to this directory.}

\item{entries}{[vector of , required] Entries, i.e. subdirectories or contained files.All entries are readable, i.e. can be used as input or navigated into.}
}
\description{
A list defining a DirectoryModel
}
