% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotaxarapi-package.R
\docType{package}
\name{ecotaxarapi-package}
\alias{ecotaxarapi}
\alias{ecotaxarapi-package}
\title{ecotaxarapi: R Package to access EcoTaxa's API}
\description{
This package exposes all objects and endpoints of EcoTaxa's API (documented at http://ecotaxa.obs-vlfr.fr/api/docs) as R functions, allowing to easily call them from R.
}
\author{
\strong{Maintainer}: Jean-Olivier Irisson \email{irisson@normalesup.org}

Authors:
\itemize{
  \item Thelma Panaiotis \email{thelma.panaiotis@imev-mer.fr}
}

}
\keyword{internal}
