% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emodnet_format_export.R
\name{emodnet_format_export}
\alias{emodnet_format_export}
\title{Emodnet Format Export}
\usage{
emodnet_format_export(
  collection_id,
  dry_run,
  with_zeroes,
  auto_morpho,
  with_computations
)
}
\arguments{
\item{collection_id}{integer; Internal, the unique numeric id of this collection.}

\item{dry_run}{boolean; If set, then only a diagnostic of doability will be done.}

\item{with_zeroes}{boolean; If set, then \emph{absent} records will be generated, in the relevant samples, for categories present in other samples.}

\item{auto_morpho}{boolean; If set, then any object classified on a Morpho category will be added to the count of the nearest Phylo parent, upward in the tree.}

\item{with_computations}{boolean; If set, then an attempt will be made to compute organisms concentrations and biovolumes.}
}
\description{
\strong{Export the collection in EMODnet format}, @see https://www.emodnet-ingestion.eu Produces a DwC-A archive into a temporary directory, ready for download. Maybe useful, a reader in Python: https://python-dwca-reader.readthedocs.io/en/latest/index.html 🔒 \emph{For admins only.}
}
