% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase_collection.R
\name{erase_collection}
\alias{erase_collection}
\title{Erase Collection}
\usage{
erase_collection(collection_id)
}
\arguments{
\item{collection_id}{integer; Internal, the unique numeric id of this collection.}
}
\description{
\strong{Delete the collection}, i.e. the precious fields, as the projects are just linked-at from the collection. Note: Only manageable collections can be deleted.
}
