% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase_job.R
\name{erase_job}
\alias{erase_job}
\title{Erase Job}
\usage{
erase_job(job_id)
}
\arguments{
\item{job_id}{integer; Internal, the unique numeric id of this job.}
}
\description{
\strong{Delete the job} from DB, with associated storage. Return \strong{NULL upon success.} If the job is running then kill it. 🔒 The job must be accessible to current user.
}
