% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_object_set.R
\name{export_object_set}
\alias{export_object_set}
\title{Export Object Set}
\usage{
export_object_set(Body_export_object_set_object_set_export_post)
}
\arguments{
\item{Body_export_object_set_object_set_export_post}{Output of Body_export_object_set_object_set_export_post();}
}
\description{
\strong{Start an export job for the given object set and options.}
}
