% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_object_set_backup.R
\name{export_object_set_backup}
\alias{export_object_set_backup}
\title{Export Object Set Backup}
\usage{
export_object_set_backup(
  Body_export_object_set_backup_object_set_export_backup_post
)
}
\arguments{
\item{Body_export_object_set_backup_object_set_export_backup_post}{Output of \code{\link[=Body_export_object_set_backup_object_set_export_backup_post]{Body_export_object_set_backup_object_set_export_backup_post()}};}
}
\description{
Start a backup export job for the given object set and options. If filters are empty, the produced zip will contain the full project. 🔒 Current user needs \emph{at least Read} right on the requested project.
}
