% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_object_set_general.R
\name{export_object_set_general}
\alias{export_object_set_general}
\title{Export Object Set General}
\usage{
export_object_set_general(
  Body_export_object_set_general_object_set_export_general_post
)
}
\arguments{
\item{Body_export_object_set_general_object_set_export_general_post}{Output of Body_export_object_set_general_object_set_export_general_post();}
}
\description{
Start a general-purpose export job for the given object set and options. 🔒 Current user needs \emph{at least Read} right on the requested project.
}
