% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_object_set_summary.R
\name{export_object_set_summary}
\alias{export_object_set_summary}
\title{Export Object Set Summary}
\usage{
export_object_set_summary(
  Body_export_object_set_summary_object_set_export_summary_post
)
}
\arguments{
\item{Body_export_object_set_summary_object_set_export_summary_post}{Output of Body_export_object_set_summary_object_set_export_summary_post();}
}
\description{
Start a summary export job for the given object set and options. 🔒 Current user needs \emph{at least Read} right on the requested project.
}
