% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportImagesOptionsEnum.R
\name{ExportImagesOptionsEnum}
\alias{ExportImagesOptionsEnum}
\title{ExportImagesOptionsEnum}
\format{
string containing all, first or none
}
\description{
A vector defining a ExportImagesOptionsEnum
}
