% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportReq.R
\name{ExportReq}
\alias{ExportReq}
\title{Export request Model}
\usage{
ExportReq(
  project_id,
  exp_type,
  use_latin1 = NULL,
  tsv_entities,
  split_by,
  coma_as_separator,
  format_dates_times,
  with_images,
  with_internal_ids,
  only_first_image,
  sum_subtotal,
  pre_mapping = NULL,
  formulae = NULL,
  out_to_ftp
)
}
\arguments{
\item{project_id}{[integer, required] The project to export.}

\item{exp_type}{[object of type \link{ExportTypeEnum}, required] The export type.}

\item{use_latin1}{[boolean] Export using latin 1 character set, AKA iso-8859-1. Default is utf-8.}

\item{tsv_entities}{[string, required] For 'TSV' type, the entities to export, one letter for each of O(bject), P(rocess), A(cquisition), S(ample), classification H(istory), C(omments).}

\item{split_by}{[string, required] For 'TSV' type, inside archives, split in one directory per... 'sample', 'taxo' or '' (no split).}

\item{coma_as_separator}{[boolean, required] For 'TSV' type, use a , instead of . for decimal separator.}

\item{format_dates_times}{[boolean, required] For 'TSV' type, format dates and times using - and : respectively.}

\item{with_images}{[boolean, required] For 'BAK' and 'DOI' types, export images as well.}

\item{with_internal_ids}{[boolean, required] For 'BAK' and 'DOI' types, export internal DB IDs.}

\item{only_first_image}{[boolean, required] For 'DOI' type, export only first (displayed) image.}

\item{sum_subtotal}{[object of type \link{SummaryExportGroupingEnum}, required] For 'SUM', 'ABO', 'CNC' and 'BIV' types, if computations should be combined. Per A(cquisition) or S(ample) or \if{html}{\out{<Empty>}}(just taxa).}

\item{pre_mapping}{[list] For 'ABO', 'CNC' and 'BIV' types types, mapping from present taxon (key) to output replacement one (value). Use a null replacement to \emph{discard} the present taxon.}

\item{formulae}{[list] Transitory: For 'CNC' and 'BIV' type, how to get values from DB free columns. Python syntax, prefixes are 'sam', 'ssm' and 'obj'.Variables used in computations are 'total_water_volume', 'subsample_coef' and 'individual_volume'}

\item{out_to_ftp}{[boolean, required] Copy result file to FTP area. Original file is still available.}
}
\description{
A list defining a Export request Model
}
