% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportRsp.R
\name{ExportRsp}
\alias{ExportRsp}
\title{ExportRsp}
\usage{
ExportRsp(errors = NULL, warnings = NULL, job_id = NULL)
}
\arguments{
\item{errors}{[vector of string] Showstopper problems found preventing building the archive.}

\item{warnings}{[vector of string] Problems found while building the archive, which do not prevent producing it.}

\item{job_id}{[integer] The created job, 0 if there were problems.}
}
\description{
A list defining a ExportRsp
}
