% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportSplitOptionsEnum.R
\name{ExportSplitOptionsEnum}
\alias{ExportSplitOptionsEnum}
\title{ExportSplitOptionsEnum}
\format{
string containing sample, acquisition, taxon or none
}
\description{
A vector defining a ExportSplitOptionsEnum
}
