% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralExportReq.R
\name{GeneralExportReq}
\alias{GeneralExportReq}
\title{General Export request Model}
\usage{
GeneralExportReq(
  collection_id = NULL,
  project_id,
  split_by = NULL,
  with_images = NULL,
  with_internal_ids = NULL,
  with_types_row = NULL,
  only_annotations = NULL,
  out_to_ftp = NULL
)
}
\arguments{
\item{collection_id}{[integer] The Collection to export if requested.}

\item{project_id}{[integer, string, required] The project(int) or projects (str, project ids list) to export.}

\item{split_by}{[object of type \link{ExportSplitOptionsEnum}] If not none, separate (in ZIP sub-directories) output per given field.}

\item{with_images}{[object of type \link{ExportImagesOptionsEnum}] Add in ZIP first (i.e. visible) image, all images, or no image.⚠️ 'all' means maybe several lines per object in TSVs.}

\item{with_internal_ids}{[boolean] Export internal database IDs.}

\item{with_types_row}{[boolean] Add an EcoTaxa-compatible second line with types.}

\item{only_annotations}{[boolean] Only save objects' last annotation data.}

\item{out_to_ftp}{[boolean] Copy result file to FTP area. Original file is still available.}
}
\description{
A list defining a General Export request Model
}
