% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_admin_users.R
\name{get_admin_users}
\alias{get_admin_users}
\title{Get Admin Users}
\usage{
get_admin_users()
}
\description{
\strong{List application administrators}, themselves being users. ğŸ”’ Any authenticated user can access the list.
}
