% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collection.R
\name{get_collection}
\alias{get_collection}
\title{Get Collection}
\usage{
get_collection(collection_id)
}
\arguments{
\item{collection_id}{integer; Internal, the unique numeric id of this collection.}
}
\description{
Returns \strong{information about the collection} corresponding to the given id. Note: The collection is returned only if manageable.
}
