% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collection_taxonomy_recast.R
\name{get_collection_taxonomy_recast}
\alias{get_collection_taxonomy_recast}
\title{Read Collection Taxo Recast}
\usage{
get_collection_taxonomy_recast(collection_id)
}
\arguments{
\item{collection_id}{integer; Internal, the unique numeric id of this collection.}
}
\description{
\strong{Read the collection taxonomy recast}. \strong{Returns NULL upon success.} Note: The collection data is returned only if manageable.
}
