% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_guest.R
\name{get_guest}
\alias{get_guest}
\title{Get Guest}
\usage{
get_guest(guest_id)
}
\arguments{
\item{guest_id}{integer; Internal, the unique numeric id of this guest.}
}
\description{
Returns \strong{information about the user} corresponding to the given id.
}
