% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_guests.R
\name{get_guests}
\alias{get_guests}
\title{Get Guests}
\usage{
get_guests(ids = NULL, fields = NULL)
}
\arguments{
\item{ids}{string; String containing the list of one or more id separated by non-num char. \strong{If several ids are provided}, one full info is returned per user.}

\item{fields}{string; Return the default fields (typically used in conjunction with an additional field list). For users list display purpose.}
}
\description{
Returns the list of \strong{all guests} with their full information, or just some of them if their ids are provided. 🔒 \emph{For admins and managers only.}
}
