% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_job.R
\name{get_job}
\alias{get_job}
\title{Get Job}
\usage{
get_job(job_id)
}
\arguments{
\item{job_id}{integer; Internal, the unique numeric id of this job.}
}
\description{
Returns \strong{information about the job} corresponding to the given id.
}
