% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_job_log_file.R
\name{get_job_log_file}
\alias{get_job_log_file}
\title{Get Job Log File}
\usage{
get_job_log_file(job_id)
}
\arguments{
\item{job_id}{integer; Internal, the unique numeric id of this job.}
}
\description{
\strong{Return the log file produced by given job.} 🔒 The job must be accessible to current user.
}
