% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_object_set.R
\name{get_object_set}
\alias{get_object_set}
\title{Get Object Set}
\usage{
get_object_set(
  project_id,
  fields = NULL,
  order_field = NULL,
  window_start = NULL,
  window_size = NULL,
  ProjectFilters
)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}

\item{fields}{string;  Specify the needed object (and ancillary entities) fields. It follows the naming convention 'prefix.field' : Prefix is either 'obj' for main object, 'fre' for free fields, 'img' for the visible image. The column obj.imgcount contains the total count of images for the object. Use a comma to separate fields. More help : You can get the field labels by parsing the classiffieldlist returned by a call to https://ecotaxa.obs-vlfr.fr/api/docs#/projects/project_query_projects__project_id__get. \strong{Note that the following fields must be prefixed with the header "obj."} (for example → obj.orig_id): acquisid classif_auto_id, classif_auto_score, classif_auto_when, classif_crossvalidation_id, classif_id, classif_qual, classif_who, classif_when, complement_info, depth_max, depth_min, latitude, longitude, objdate, object_link, objid, objtime, orig_id, random_value, similarity, sunpos. \strong{Note that the following fields must be prefixed with the header "img."} (for example → img.file_name): file_name, height, imgid, imgrank, file_name, objid, orig_file_name, thumb_file_name, thumb_height, thumb_width, width. \strong{Note that the following fields must be prefixed with the header "txo."} (for example → txo.display_name): creation_datetime, creator_email, display_name, id, id_instance, id_source, lastupdate_datetime, name, nbrobj, nbrobjcum, parent_id, rename_to, source_desc, source_url, taxostatus, taxotype. \strong{All other fields must be prefixed by the header "fre."} (for example → fre.circ.).}

\item{order_field}{string; Order the result using given field. If prefixed with "-" then it will be reversed. When using \emph{special syntax ss-Innnn}, the order is similarity with given (by its ID) object.}

\item{window_start}{integer;  Allows to return only a slice of the result, by skipping window_start objects before returning data. If no \strong{unique order} is specified, the result can vary for same call and conditions.}

\item{window_size}{integer;  Allows to return only a slice of the result, by returning a \emph{maximum} of window_size lines. If no \strong{unique order} is specified, the result can vary for same call and conditions.}

\item{ProjectFilters}{Output of \code{\link[=ProjectFilters]{ProjectFilters()}}; How to reduce project data.}
}
\description{
Returns \strong{filtered object Ids} for the given project.
}
