% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_object_set_similarity_search.R
\name{get_object_set_similarity_search}
\alias{get_object_set_similarity_search}
\title{Object Similarity Search}
\usage{
get_object_set_similarity_search(
  project_id,
  object_id,
  size = NULL,
  ProjectFilters
)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project to search in.}

\item{object_id}{integer; Object ID to search similar for.}

\item{size}{integer; Return at maximum this number of object IDs, by default 100.}

\item{ProjectFilters}{Output of \code{\link[=ProjectFilters]{ProjectFilters()}}; How to reduce project data.}
}
\description{
Returns, in given project, the objects matching the filter and similar to the queried one.
}
