% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_object_set_summary.R
\name{get_object_set_summary}
\alias{get_object_set_summary}
\title{Get Object Set Summary}
\usage{
get_object_set_summary(project_id, only_total, ProjectFilters)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}

\item{only_total}{boolean; If True, returns only the \strong{Total number of objects}. Else returns also the \strong{Number of validated ones}, the \strong{number of Dubious ones} and the number of \strong{predicted ones}.}

\item{ProjectFilters}{Output of \code{\link[=ProjectFilters]{ProjectFilters()}}; How to reduce project data.}
}
\description{
For the given project, with given filters, \strong{return the classification summary}. i.e.: - Total number of objects And optionally - Number of Validated ones - Number of Dubious ones - Number of Predicted ones
}
