% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_organizations.R
\name{get_organizations}
\alias{get_organizations}
\title{Get Organizations}
\usage{
get_organizations(ids = NULL)
}
\arguments{
\item{ids}{string; String containing the list of one or more id separated by non-num char. \strong{If several ids are provided}, one full info is returned per user.}
}
\description{
\strong{Search for organizations.}
}
