% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxon_in_central.R
\name{get_taxon_in_central}
\alias{get_taxon_in_central}
\title{Get Taxon In Central}
\usage{
get_taxon_in_central(taxon_id)
}
\arguments{
\item{taxon_id}{integer; Internal, the unique numeric id of this taxon.}
}
\description{
Return \strong{EcoTaxoServer full record for this taxon}.
}
