% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_user}
\alias{get_user}
\title{Get User}
\usage{
get_user(user_id)
}
\arguments{
\item{user_id}{integer; Internal, the unique numeric id of this user.}
}
\description{
Returns \strong{information about the user} corresponding to the given id.
}
