% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuestModel.R
\name{GuestModel}
\alias{GuestModel}
\title{GuestModel}
\usage{
GuestModel(id, email, name, country = NULL, orcid = NULL, organisation)
}
\arguments{
\item{id}{[integer, required] The unique numeric id of this user.}

\item{email}{[string, required] User's email address, as text, used during registration.}

\item{name}{[string, required] User's full name, as text.}

\item{country}{[string] The country name, as text (but chosen in a consistent list).}

\item{orcid}{[string] The orcid id https://support.orcid.org.}

\item{organisation}{[string, required] User's organisation name, as text.}
}
\description{
A list defining a GuestModel
}
