% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistoricalClassification.R
\name{HistoricalClassification}
\alias{HistoricalClassification}
\title{HistoricalClassification}
\usage{
HistoricalClassification(
  objid = NULL,
  classif_id = NULL,
  classif_date = NULL,
  classif_who = NULL,
  classif_type = NULL,
  classif_qual = NULL,
  classif_score = NULL,
  user_name = NULL,
  taxon_name = NULL
)
}
\arguments{
\item{objid}{[integer] The object Id.}

\item{classif_id}{[integer] The classification Id.}

\item{classif_date}{[string] The classification date.}

\item{classif_who}{[integer] The user who manualy classify this object.}

\item{classif_type}{[string] The type of classification. Could be \strong{A} for Automatic or \strong{M} for Manual.}

\item{classif_qual}{[string] The classification qualification. Could be \strong{P} for predicted, \strong{V} for validated or \strong{D} for Dubious.}

\item{classif_score}{[number] The classification score is generally between 0 and 1. This is a confidence score, in the fact that, the taxon prediction for this object is correct.}

\item{user_name}{[string] The name of the user who classified this object.}

\item{taxon_name}{[string] The taxon name of the object.}
}
\description{
A list defining a HistoricalClassification
}
