% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistoricalLastClassif.R
\name{HistoricalLastClassif}
\alias{HistoricalLastClassif}
\title{HistoricalLastClassif}
\usage{
HistoricalLastClassif(
  objid = NULL,
  classif_id = NULL,
  histo_classif_date = NULL,
  histo_classif_id = NULL,
  histo_classif_type = NULL,
  histo_classif_qual = NULL,
  histo_classif_who = NULL,
  histo_classif_score = NULL
)
}
\arguments{
\item{objid}{[integer] The object Id.}

\item{classif_id}{[integer] The classification Id.}

\item{histo_classif_date}{[string] The classification date.}

\item{histo_classif_id}{[integer] The classification Id.}

\item{histo_classif_type}{[string] The type of classification. Could be \strong{A} for Automatic or \strong{M} for Manual.}

\item{histo_classif_qual}{[string] The classification qualification. Could be \strong{P} for predicted, \strong{V} for validated or \strong{D} for Dubious.}

\item{histo_classif_who}{[integer] The user who manually classified this object.}

\item{histo_classif_score}{[number] Histo Classif Score}
}
\description{
A list defining a HistoricalLastClassif
}
