% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPValidationError.R
\name{HTTPValidationError}
\alias{HTTPValidationError}
\title{HTTPValidationError}
\usage{
HTTPValidationError(detail = NULL)
}
\arguments{
\item{detail}{[vector of ] Detail}
}
\description{
A list defining a HTTPValidationError
}
