% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_file.R
\name{import_file}
\alias{import_file}
\title{Import File}
\usage{
import_file(
  project_id,
  source_path = NULL,
  taxo_mappings = NULL,
  skip_loaded_files = NULL,
  skip_existing_objects = NULL,
  update_mode = NULL
)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}

\item{source_path}{string; Source path on server, to zip or plain directory. The path can be returned by a file upload (absolute), otherwise it's relative to shared file area root.}

\item{taxo_mappings}{object; Optional taxonomy mapping, the key specifies the taxonomy ID found in file and the value specifies the final taxonomy ID to write.}

\item{skip_loaded_files}{boolean; If true skip loaded files, else don't.}

\item{skip_existing_objects}{boolean; If true skip existing objects, else don't.}

\item{update_mode}{string; Update data ('Yes'), including classification ('Cla').}
}
\description{
\strong{Validate or do a real import} of an EcoTaxa archive or directory.
}
