% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportReq.R
\name{ImportReq}
\alias{ImportReq}
\title{Import request Model}
\usage{
ImportReq(
  source_path,
  taxo_mappings = NULL,
  skip_loaded_files = NULL,
  skip_existing_objects = NULL,
  update_mode = NULL
)
}
\arguments{
\item{source_path}{[string, required] Source path on server, to zip or plain directory.     The path can be returned by a file upload (absolute),     otherwise it's relative to shared file area root.}

\item{taxo_mappings}{[list] Optional taxonomy mapping, the key specifies the taxonomy ID found in file and the value specifies the final taxonomy ID to write.}

\item{skip_loaded_files}{[boolean] If true skip loaded files, else don't.}

\item{skip_existing_objects}{[boolean] If true skip existing objects, else don't.}

\item{update_mode}{[string] Update data ('Yes'), including classification ('Cla').}
}
\description{
A list defining a Import request Model
}
