% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportRsp.R
\name{ImportRsp}
\alias{ImportRsp}
\title{ImportRsp}
\usage{
ImportRsp(job_id, errors = NULL)
}
\arguments{
\item{job_id}{[integer, required] The job which was created for the run.}

\item{errors}{[vector of string] Errors from analysis.}
}
\description{
A list defining a ImportRsp
}
