% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_query.R
\name{instrument_query}
\alias{instrument_query}
\title{Instrument Query}
\usage{
instrument_query(project_ids)
}
\arguments{
\item{project_ids}{string; String containing the list of one or more project ids, separated by non-num char, or 'all' for all instruments.}
}
\description{
Returns the list of instruments, inside specific project(s) or globally.
}
