% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobModel.R
\name{JobModel}
\alias{JobModel}
\title{JobModel}
\usage{
JobModel(
  params = NULL,
  result = NULL,
  errors = NULL,
  question = NULL,
  reply = NULL,
  inside = NULL,
  id,
  owner_id,
  type,
  state = NULL,
  step = NULL,
  progress_pct = NULL,
  progress_msg = NULL,
  creation_date,
  updated_on
)
}
\arguments{
\item{params}{[list] Creation parameters.}

\item{result}{[list] Final result of the run.}

\item{errors}{[vector of string] The errors seen during last step.}

\item{question}{[list] The data provoking job move to Asking state.}

\item{reply}{[list] The data provided as a reply to the question.}

\item{inside}{[list] Internal state of the job.}

\item{id}{[integer, required] Job unique identifier.}

\item{owner_id}{[integer, required] The user who created and thus owns the job.}

\item{type}{[string, required] The job type, e.g. import, export...}

\item{state}{[string] What the job is doing. Could be 'P' for Pending (Waiting for an execution thread), 'R' for Running (Being executed inside a thread), 'A' for Asking (Needing user information before resuming), 'E' for Error (Stopped with error), 'F' for Finished (Done).}

\item{step}{[integer] Where in the workflow the job is.}

\item{progress_pct}{[integer] The progress percentage for UI.}

\item{progress_msg}{[string] The message for UI, short version.}

\item{creation_date}{[string, required] The date of creation of the Job, as text formatted according to the ISO 8601 standard.}

\item{updated_on}{[string, required] Last time that anything changed in present line.}
}
\description{
A list defining a JobModel
}
