% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_collections.R
\name{list_collections}
\alias{list_collections}
\title{List Collections}
\usage{
list_collections(collection_ids = NULL, fields = NULL)
}
\arguments{
\item{collection_ids}{string; limit the list to a set of ids.}

\item{fields}{string; Return the default fields (typically used in conjunction with an additional field list). For users list display purpose.}
}
\description{
\strong{Search for collections.} Note: Only collections where the current user is manager are returned. All collections are returned if the user is application administrator
}
