% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_common_files.R
\name{list_common_files}
\alias{list_common_files}
\title{List Common Files}
\usage{
list_common_files(path)
}
\arguments{
\item{path}{string; NULL}
}
\description{
\strong{List the common files} which are usable for some file-related operations. \emph{e.g. import.}
}
