% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_jobs.R
\name{list_jobs}
\alias{list_jobs}
\title{List Jobs}
\usage{
list_jobs(for_admin)
}
\arguments{
\item{for_admin}{boolean; If FALSE return the jobs for current user, else return all of them.}
}
\description{
\strong{Return the jobs} for current user, or all of them if admin is asked for.
}
