% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_my_files.R
\name{list_my_files}
\alias{list_my_files}
\title{List My Files}
\usage{
list_my_files(sub_path)
}
\arguments{
\item{sub_path}{string; NULL}
}
\description{
\strong{List the private files} from user files directory which are usable for some file-related operations. A sub_path starting with "/" is considered relative to user folder. \emph{e.g. import.}
}
