% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_projects.R
\name{list_projects}
\alias{list_projects}
\title{List Projects}
\usage{
list_projects(
  project_ids = NULL,
  not_granted = NULL,
  for_managing = NULL,
  order_field = NULL,
  fields = NULL,
  window_start = NULL,
  window_size = NULL
)
}
\arguments{
\item{project_ids}{string; Limit the list to a set of ids.}

\item{not_granted}{boolean; Return projects on which the current user has \emph{no permission}, but visible to him/her.}

\item{for_managing}{boolean; Return projects that can be written to (including erased) by the current user.}

\item{order_field}{string; One of \link{'projid', 'title', 'access', 'status', 'objcount', 'pctvalidated', 'pctclassified', 'classifsettings', 'classiffieldlist', 'popoverfieldlist', 'comments', 'rf_models_used', 'cnn_network_id', 'formulae', 'instrument', 'instrument_url', 'highest_right'}}

\item{fields}{string; Return the default fields (typically used in conjunction with an additional field list). For users list display purpose.}

\item{window_start}{integer; Skip \code{window_start} before returning data.}

\item{window_size}{integer; Return only \code{window_size} lines.}
}
\description{
Returns \strong{projects which the current user has explicit permission to access, with fields options.} Note that, for performance reasons, in returned ProjectModels, field 'highest_rank' is NOT valued (unlike in simple query). The same information can be found in 'managers', 'annotators' and 'viewers' lists.
}
