% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_user_files.R
\name{list_user_files}
\alias{list_user_files}
\title{List User Files}
\usage{
list_user_files(sub_path)
}
\arguments{
\item{sub_path}{string; NULL}
}
\description{
\strong{List the private files} which are usable for some file-related operations. A sub_path starting with "/" is considered relative to user folder. \emph{e.g. import.}
}
