% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login}
\usage{
login(LoginReq)
}
\arguments{
\item{LoginReq}{Output of \code{\link[=LoginReq]{LoginReq()}};}
}
\description{
\strong{Login barrier,} If successful, the login will return a \strong{JWT} which will have to be used in bearer authentication scheme for subsequent calls.
}
