% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoginReq.R
\name{LoginReq}
\alias{LoginReq}
\title{Login request Model}
\usage{
LoginReq(password, username)
}
\arguments{
\item{password}{[string, required] User password.}

\item{username}{[string, required] User email used during registration.}
}
\description{
A list defining a Login request Model
}
