% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeRsp.R
\name{MergeRsp}
\alias{MergeRsp}
\title{MergeRsp}
\usage{
MergeRsp(errors = NULL)
}
\arguments{
\item{errors}{[vector of string] The errors found during processing.}
}
\description{
A list defining a MergeRsp
}
