% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinimalCollectionBO.R
\name{MinimalCollectionBO}
\alias{MinimalCollectionBO}
\title{MinimalCollectionBO}
\usage{
MinimalCollectionBO(
  id,
  external_id = NULL,
  title,
  short_title = NULL,
  provider_user,
  contact_user = NULL,
  project_ids
)
}
\arguments{
\item{id}{[integer, required] Id}

\item{external_id}{[string] External Id}

\item{title}{[string, required] Title}

\item{short_title}{[string] Short Title}

\item{provider_user}{[integer, required] Provider User}

\item{contact_user}{[integer] Contact User}

\item{project_ids}{[vector of integer, required] Project Ids}
}
\description{
A list defining a MinimalCollectionBO
}
