% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinimalUserBO.R
\name{MinimalUserBO}
\alias{MinimalUserBO}
\title{MinimalUserBO}
\usage{
MinimalUserBO(id = NULL, name = NULL)
}
\arguments{
\item{id}{[integer] Id}

\item{name}{[string] Name}
}
\description{
A list defining a MinimalUserBO
}
