% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinUserModel.R
\name{MinUserModel}
\alias{MinUserModel}
\title{MinUserModel}
\usage{
MinUserModel(id, email, name, organisation)
}
\arguments{
\item{id}{[integer, required] The unique numeric id of this user.}

\item{email}{[string, required] User's email address, as text, used during registration.}

\item{name}{[string, required] User's full name, as text.}

\item{organisation}{[string, required] User's organisation name, as text.}
}
\description{
A list defining a MinUserModel
}
