% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_my_file.R
\name{move_my_file}
\alias{move_my_file}
\title{Move File}
\usage{
move_my_file()
}
\description{
\strong{Move (or rename depending on source and dest path) a file or directory in the current user files directory.} The returned text will contain a server-side path which is usable for some file-related operations.
}
