% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_query.R
\name{object_query}
\alias{object_query}
\title{Object Query}
\usage{
object_query(object_id)
}
\arguments{
\item{object_id}{integer; Internal, the unique numeric id of this object.}
}
\description{
Returns \strong{information about the object} corresponding to the given id. 🔒 Anonymous reader can do if the project has the right rights :)
}
