% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_query_history.R
\name{object_query_history}
\alias{object_query_history}
\title{Object Query History}
\usage{
object_query_history(object_id)
}
\arguments{
\item{object_id}{integer; Internal, the unique numeric id of this object.}
}
\description{
Returns \strong{information about the object's history} corresponding to the given id.
}
