% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_set_predictions.R
\name{object_set_predictions}
\alias{object_set_predictions}
\title{Query Object Set Predictions}
\usage{
object_set_predictions(object_ids_list)
}
\arguments{
\item{object_ids_list}{array; The list of object ids.}
}
\description{
** Return last prediction information for a set of objects, by their IDs.
}
