% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectHeaderModel.R
\name{ObjectHeaderModel}
\alias{ObjectHeaderModel}
\title{ObjectHeaderModel}
\usage{
ObjectHeaderModel(
  objid,
  acquisid,
  classif_id = NULL,
  objtime = NULL,
  latitude = NULL,
  longitude = NULL,
  depth_min = NULL,
  depth_max = NULL,
  objdate = NULL,
  classif_qual = NULL,
  sunpos = NULL,
  classif_when = NULL,
  classif_who = NULL,
  classif_auto_id = NULL,
  classif_auto_when = NULL,
  classif_auto_score = NULL,
  orig_id,
  object_link = NULL,
  complement_info = NULL
)
}
\arguments{
\item{objid}{[integer, required] The object Id.}

\item{acquisid}{[integer, required] The parent acquisition Id.}

\item{classif_id}{[integer] The classification Id.}

\item{objtime}{[string] Object time}

\item{latitude}{[number] The latitude.}

\item{longitude}{[number] The longitude.}

\item{depth_min}{[number] The min depth.}

\item{depth_max}{[number] The min depth.}

\item{objdate}{[string] Object date}

\item{classif_qual}{[string] The classification qualification. Could be \strong{P} for predicted, \strong{V} for validated or \strong{D} for Dubious.}

\item{sunpos}{[string] Sun position, from date, time and coords.}

\item{classif_when}{[string] The classification date.}

\item{classif_who}{[integer] The user who manualy classify this object.}

\item{classif_auto_id}{[integer] Set if the object was ever predicted, remain forever with these value. Reflect the 'last state' only if classif_qual is 'P'.}

\item{classif_auto_when}{[string] Set if the object was ever predicted, remain forever with these value. Reflect the 'last state' only if classif_qual is 'P'. The classification date.}

\item{classif_auto_score}{[number] Set if the object was ever predicted, remain forever with these value. Reflect the 'last state' only if classif_qual is 'P'. The classification auto score is generally between 0 and 1. This is a confidence score, in the fact that, the taxon prediction for this object is correct.}

\item{orig_id}{[string, required] Original object ID from initial TSV load.}

\item{object_link}{[string] Object link.}

\item{complement_info}{[string] Complement info}
}
\description{
A list defining a ObjectHeaderModel
}
