% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectSetQueryRsp.R
\name{ObjectSetQueryRsp}
\alias{ObjectSetQueryRsp}
\title{ObjectSetQueryRsp}
\usage{
ObjectSetQueryRsp(
  object_ids = NULL,
  acquisition_ids = NULL,
  sample_ids = NULL,
  project_ids = NULL,
  details = NULL,
  total_ids = NULL
)
}
\arguments{
\item{object_ids}{[vector of integer] Matching object IDs.}

\item{acquisition_ids}{[vector of integer] Parent (acquisition) IDs.}

\item{sample_ids}{[vector of integer] Parent (sample) IDs.}

\item{project_ids}{[vector of integer] Project Ids.}

\item{details}{[vector of array] Requested fields, in request order.}

\item{total_ids}{[integer] Total rows returned by the query, even if it was window-ed.}
}
\description{
A list defining a ObjectSetQueryRsp
}
