% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectSetRevertToHistoryRsp.R
\name{ObjectSetRevertToHistoryRsp}
\alias{ObjectSetRevertToHistoryRsp}
\title{ObjectSetRevertToHistoryRsp}
\usage{
ObjectSetRevertToHistoryRsp(last_entries = NULL, classif_info = NULL)
}
\arguments{
\item{last_entries}{[vector of ] Object + last classification}

\item{classif_info}{[list] Classification names (self+parent) for involved IDs.}
}
\description{
A list defining a ObjectSetRevertToHistoryRsp
}
