% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectSetSummaryRsp.R
\name{ObjectSetSummaryRsp}
\alias{ObjectSetSummaryRsp}
\title{ObjectSetSummaryRsp}
\usage{
ObjectSetSummaryRsp(
  total_objects = NULL,
  validated_objects = NULL,
  dubious_objects = NULL,
  predicted_objects = NULL
)
}
\arguments{
\item{total_objects}{[integer] Total number of objects in the set.}

\item{validated_objects}{[integer] Number of validated objects in the set.}

\item{dubious_objects}{[integer] Number of dubious objects in the set.}

\item{predicted_objects}{[integer] Number of predicted objects in the set.}
}
\description{
A list defining a ObjectSetSummaryRsp
}
