% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganizationModel.R
\name{OrganizationModel}
\alias{OrganizationModel}
\title{OrganizationModel}
\usage{
OrganizationModel(id, name, directories = NULL)
}
\arguments{
\item{id}{[integer, required] OrganizationIDT unique identifier.}

\item{name}{[string, required] Organization's full name, as text.}

\item{directories}{[string] References to official directories where the organization is referenced, separated by ,.}
}
\description{
A list defining a OrganizationModel
}
