% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_user_file.R
\name{post_user_file}
\alias{post_user_file}
\title{Put User File}
\usage{
post_user_file(Body_put_user_file_user_files__post)
}
\arguments{
\item{Body_put_user_file_user_files__post}{Output of \code{\link[=Body_put_user_file_user_files__post]{Body_put_user_file_user_files__post()}};}
}
\description{
\strong{Upload a file for the current user files directory.} The returned text will contain a server-side path which is usable for some file-related operations. \emph{e.g. import.}
}
