% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionInfoRsp.R
\name{PredictionInfoRsp}
\alias{PredictionInfoRsp}
\title{PredictionInfoRsp}
\usage{
PredictionInfoRsp(result = NULL)
}
\arguments{
\item{result}{[vector of array] List of lists \link{object ID, category ID, score for category}.}
}
\description{
A list defining a PredictionInfoRsp
}
