% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionInfoT.R
\name{PredictionInfoT}
\alias{PredictionInfoT}
\title{PredictionInfoT}
\usage{
PredictionInfoT(object_id, classif_id, score)
}
\arguments{
\item{object_id}{[integer, required] Object Id}

\item{classif_id}{[integer, required] Classif Id}

\item{score}{[number, required] Score}
}
\description{
A list defining a PredictionInfoT
}
