% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionReq.R
\name{PredictionReq}
\alias{PredictionReq}
\title{Prediction Request}
\usage{
PredictionReq(
  project_id,
  source_project_ids,
  learning_limit = NULL,
  features,
  categories,
  use_scn = NULL,
  pre_mapping
)
}
\arguments{
\item{project_id}{[integer, required] The destination project, of which objects will be predicted.}

\item{source_project_ids}{[vector of integer, required] The source projects, objects in them will serve as reference.}

\item{learning_limit}{[integer] When set (to a positive value), there will be this number  of objects, \emph{per category}, in the learning set.}

\item{features}{[vector of string, required] The object features AKA free column, to use in the algorithm. Features must be common to all projects, source ones and destination one.}

\item{categories}{[vector of integer, required] In source projects, only objects validated with these categories will be considered.}

\item{use_scn}{[boolean] Use extra features, generated using the image, for improving the prediction.}

\item{pre_mapping}{[list, required] Categories in keys become value one before launching the ML algorithm. Any unknown value is ignored.}
}
\description{
A list defining a Prediction Request
}
