% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionRsp.R
\name{PredictionRsp}
\alias{PredictionRsp}
\title{PredictionRsp}
\usage{
PredictionRsp(errors = NULL, warnings = NULL, job_id = NULL)
}
\arguments{
\item{errors}{[vector of string] Showstopper problems found while preparing the prediction.}

\item{warnings}{[vector of string] Problems found while preparing the prediction.}

\item{job_id}{[integer] The created job, 0 if there were problems.}
}
\description{
A list defining a PredictionRsp
}
