% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_query.R
\name{process_query}
\alias{process_query}
\title{Process Query}
\usage{
process_query(process_id)
}
\arguments{
\item{process_id}{integer; Internal, the unique numeric id of this process.}
}
\description{
Returns \strong{information about the process} corresponding to the given id.
}
