% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessModel.R
\name{ProcessModel}
\alias{ProcessModel}
\title{ProcessModel}
\usage{
ProcessModel(processid, orig_id, free_columns = NULL)
}
\arguments{
\item{processid}{[integer, required] The process Id.}

\item{orig_id}{[string, required] Original process ID from initial TSV load.}

\item{free_columns}{[list] Free columns from process mapping in project}
}
\description{
A list defining a ProcessModel
}
