% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_check.R
\name{project_check}
\alias{project_check}
\title{Project Check}
\usage{
project_check(project_id)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}
}
\description{
\strong{Check consistency of a project}. With time and bugs, some consistency problems could be introduced in projects. This service aims at listing them.
}
