% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_collections.R
\name{project_collections}
\alias{project_collections}
\title{Project Collections}
\usage{
project_collections(project_id)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}
}
\description{
**Returns project collections list of id, title
}
