% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_merge.R
\name{project_merge}
\alias{project_merge}
\title{Project Merge}
\usage{
project_merge(project_id, source_project_id, dry_run)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}

\item{source_project_id}{integer; Id of the other project. All objects from this source project will be moved to the project_id above and the source project itself will be deleted.}

\item{dry_run}{boolean; If set, then only a diagnostic of doability will be done.}
}
\description{
\strong{Merge another project into this one.} It's more a phagocytosis than a merge, as all objects from this source project will be moved to the project_id above and the source project itself will be deleted. TODO: Explain a bit with it might fail (too many free columns, unique orig_ids collision)
}
