% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_query.R
\name{project_query}
\alias{project_query}
\title{Project Query}
\usage{
project_query(project_id, for_managing = NULL)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}

\item{for_managing}{boolean; For managing this project.}
}
\description{
\strong{Returns project} if it exists for current user, eventually for managing it.
}
