% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_recompute_geography.R
\name{project_recompute_geography}
\alias{project_recompute_geography}
\title{Project Recompute Geography}
\usage{
project_recompute_geography(project_id)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}
}
\description{
\strong{Recompute geography information} for all samples in project. \strong{Returns NULL upon success.} 🔒 The user has to be \emph{project manager}.
}
