% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_recompute_sunpos.R
\name{project_recompute_sunpos}
\alias{project_recompute_sunpos}
\title{Project Recompute Sunpos}
\usage{
project_recompute_sunpos(project_id)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}
}
\description{
\strong{Recompute sun position field} for all objects in project. \strong{Returns NULL upon success.} 🔒 The user has to be \emph{project manager} on the referenced project.
}
