% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_set_get_column_stats.R
\name{project_set_get_column_stats}
\alias{project_set_get_column_stats}
\title{Project Set Get Column Stats}
\usage{
project_set_get_column_stats(ids, names, limit = NULL, categories = NULL)
}
\arguments{
\item{ids}{string; String containing the list of one or more id separated by non-num char.}

\item{names}{string; Coma-separated prefixed columns, on which stats are needed.}

\item{limit}{integer; Only compute stats on this number of objects per category.}

\item{categories}{string; String containing the Categories, one or more id separated by non-num char.}
}
\description{
\strong{Returns projects validated data statistics}, for all named columns, in all given projects. The free columns here are named by the alias e.g. 'area', not technical name e.g. 'n43'. This allows getting stats on projects with different mappings, but common names.
}
