% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_set_get_projects.R
\name{project_set_get_projects}
\alias{project_set_get_projects}
\title{Project Set Get Projects}
\usage{
project_set_get_projects(ids, fields = NULL)
}
\arguments{
\item{ids}{string; String containing the list of one or more project id separated by non-num char. \strong{If several ids are provided}, one stat record will be returned per project.}

\item{fields}{string; Return the default fields (typically used in conjunction with an additional field list). To return selected fields.}
}
\description{
\strong{Returns projects statistics}, i.e. used taxa and classification states.
}
