% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_set_get_stats.R
\name{project_set_get_stats}
\alias{project_set_get_stats}
\title{Project Set Get Stats}
\usage{
project_set_get_stats(ids, taxa_ids = NULL)
}
\arguments{
\item{ids}{string; String containing the list of one or more project id separated by non-num char. \strong{If several ids are provided}, one stat record will be returned per project.}

\item{taxa_ids}{string; \strong{If several taxa_ids are provided}, one stat record will be returned per requested taxa, if populated. \strong{If taxa_ids is all}, all valued taxa in the project(s) are returned.}
}
\description{
\strong{Returns projects statistics}, i.e. used taxa and classification states.
}
