% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_stats.R
\name{project_stats}
\alias{project_stats}
\title{Project Stats}
\usage{
project_stats(project_id)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}
}
\description{
\strong{Returns stats} for a project. These stats will be returned as a list containing at index : - 0 : The \strong{title} of the project, - 1 : A string containing all \strong{freecols name and related column number}, - 2 : \strong{"(0):"} - 3 : \strong{"Total: 0 values, dup 0 values"} Then for each acquisition a pair of strings will be added to the list : - A string containing the \strong{acquisition origin id} (the \strong{number of objects for this acquisition}) : and then \strong{small stats for an acquisition of a free column values inside} : \link{ min of values ; max of values ; distribution of the different values ; mode, i.e. freq of most frequent value} - A string containing the \strong{number of total values} and the \strong{number of duplicates values} "Total: ... values, dup ... values"
}
