% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_subset.R
\name{project_subset}
\alias{project_subset}
\title{Project Subset}
\usage{
project_subset(project_id, SubsetReq)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}

\item{SubsetReq}{Output of SubsetReq(); Subset request.}
}
\description{
\strong{Subset a project into another one.}
}
