% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectColumnsModel.R
\name{ProjectColumnsModel}
\alias{ProjectColumnsModel}
\title{ProjectColumnsModel}
\usage{
ProjectColumnsModel(projid = NULL, columns = NULL, values = NULL)
}
\arguments{
\item{projid}{[integer] Project ID from the call.}

\item{columns}{[vector of string] Column names from the call.}

\item{values}{[vector of string] All rows regardless of emptiness.}
}
\description{
A list defining a ProjectColumnsModel
}
