% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectFilters.R
\name{ProjectFilters}
\alias{ProjectFilters}
\title{Project filters Model}
\usage{
ProjectFilters(
  taxo = NULL,
  taxochild = NULL,
  statusfilter = NULL,
  MapN = NULL,
  MapW = NULL,
  MapE = NULL,
  MapS = NULL,
  depthmin = NULL,
  depthmax = NULL,
  samples = NULL,
  instrum = NULL,
  daytime = NULL,
  month = NULL,
  fromdate = NULL,
  todate = NULL,
  fromtime = NULL,
  totime = NULL,
  inverttime = NULL,
  validfromdate = NULL,
  validtodate = NULL,
  freenum = NULL,
  freenumst = NULL,
  freenumend = NULL,
  freetxt = NULL,
  freetxtval = NULL,
  filt_annot = NULL,
  filt_last_annot = NULL
)
}
\arguments{
\item{taxo}{[string] Coma-separated list of numeric taxonomy/category ids. Only include objects classified with one of them.}

\item{taxochild}{[string] If 'Y' and taxo is set, also include children of each member of 'taxo' list in taxonomy tree.}

\item{statusfilter}{[string] Include objects with given status:             'NV': Not validated              'PV': Predicted or Validated              'PVD': Predicted or Validated or Dubious             'NVM': Validated, but not by me              'VM': Validated by me              'U': Not classified             other: direct equality comparison with DB value}

\item{MapN}{[string] If all 4 are set (MapN, MapW, MapE, MapS), include objects inside the defined bounding rectangle.}

\item{MapW}{[string] If all 4 are set (MapN, MapW, MapE, MapS), include objects inside the defined bounding rectangle.}

\item{MapE}{[string] If all 4 are set (MapN, MapW, MapE, MapS), include objects inside the defined bounding rectangle.}

\item{MapS}{[string] If all 4 are set (MapN, MapW, MapE, MapS), include objects inside the defined bounding rectangle.}

\item{depthmin}{[string] Positive values. If both are set (depthmin, depthmax), include objects for which both depths (min and max) are inside the range.}

\item{depthmax}{[string] Positive values. If both are set (depthmin, depthmax), include objects for which both depths (min and max) are inside the range.}

\item{samples}{[string] Coma-separated list of sample IDs, include only objects for these samples.}

\item{instrum}{[string] Instrument name, include objects for which sampling was done using this instrument.}

\item{daytime}{[string] Coma-separated list of sun position values: D for Day, U for Dusk, N for Night, A for Dawn (Aube in French).}

\item{month}{[string] Coma-separated list of month numbers, 1=Jan and so on.}

\item{fromdate}{[string] Format is 'YYYY-MM-DD', include objects collected after this date.}

\item{todate}{[string] Format is 'YYYY-MM-DD', include objects collected before this date.}

\item{fromtime}{[string] Format is 'HH24:MM:SS', include objects collected after this time of day.}

\item{totime}{[string] Format is 'HH24:MM:SS', include objects collected before this time of day.}

\item{inverttime}{[string] If '1', include objects outside fromtime and totime range.}

\item{validfromdate}{[string] Format is 'YYYY-MM-DD HH24:MI', include objects validated/set to dubious after this date+time.}

\item{validtodate}{[string] Format is 'YYYY-MM-DD HH24:MI', include objects validated/set to dubious before this date+time.}

\item{freenum}{[string] Numerical DB column number in Object as basis for the 2 following criteria (freenumst, freenumend).}

\item{freenumst}{[string] Start of included range for the column defined by freenum, in which objects are included.}

\item{freenumend}{[string] End of included range for the column defined by freenum, in which objects are included.}

\item{freetxt}{[string]  Textual DB column number as basis for following criteria (freetxtval)             If starts with 's' then it's a text column in Sample             If starts with 'a' then it's a text column in Acquisition              If starts with 'p' then it's a text column in Process              If starts with 'o' then it's a text column in Object .}

\item{freetxtval}{[string] Text to match in the column defined by freetxt, for an object to be include.}

\item{filt_annot}{[string] Coma-separated list of annotators, i.e. persons who validated the classification at any point in time.}

\item{filt_last_annot}{[string] Coma-separated list of annotators, i.e. persons who validated the classification in last.}
}
\description{
A list defining a Project filters Model
}
