% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectFiltersDict.R
\name{ProjectFiltersDict}
\alias{ProjectFiltersDict}
\title{Subset request Model}
\usage{
ProjectFiltersDict(
  taxo = NULL,
  taxochild = NULL,
  statusfilter = NULL,
  MapN = NULL,
  MapW = NULL,
  MapE = NULL,
  MapS = NULL,
  depthmin = NULL,
  depthmax = NULL,
  samples = NULL,
  instrum = NULL,
  daytime = NULL,
  month = NULL,
  fromdate = NULL,
  todate = NULL,
  fromtime = NULL,
  totime = NULL,
  inverttime = NULL,
  validfromdate = NULL,
  validtodate = NULL,
  freenum = NULL,
  freenumst = NULL,
  freenumend = NULL,
  freetxt = NULL,
  freetxtval = NULL,
  filt_annot = NULL,
  filt_last_annot = NULL
)
}
\arguments{
\item{taxo}{[string] Taxo}

\item{taxochild}{[string] Taxochild}

\item{statusfilter}{[string] Statusfilter}

\item{MapN}{[string] Mapn}

\item{MapW}{[string] Mapw}

\item{MapE}{[string] Mape}

\item{MapS}{[string] Maps}

\item{depthmin}{[string] Depthmin}

\item{depthmax}{[string] Depthmax}

\item{samples}{[string] Samples}

\item{instrum}{[string] Instrum}

\item{daytime}{[string] Daytime}

\item{month}{[string] Month}

\item{fromdate}{[string] Fromdate}

\item{todate}{[string] Todate}

\item{fromtime}{[string] Fromtime}

\item{totime}{[string] Totime}

\item{inverttime}{[string] Inverttime}

\item{validfromdate}{[string] Validfromdate}

\item{validtodate}{[string] Validtodate}

\item{freenum}{[string] Freenum}

\item{freenumst}{[string] Freenumst}

\item{freenumend}{[string] Freenumend}

\item{freetxt}{[string] Freetxt}

\item{freetxtval}{[string] Freetxtval}

\item{filt_annot}{[string] Filt Annot}

\item{filt_last_annot}{[string] Filt Last Annot}
}
\description{
A list defining a Subset request Model
}
