% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectSetColumnStatsModel.R
\name{ProjectSetColumnStatsModel}
\alias{ProjectSetColumnStatsModel}
\title{ProjectSetColumnStatsModel}
\usage{
ProjectSetColumnStatsModel(
  proj_ids = NULL,
  total = NULL,
  columns = NULL,
  counts = NULL,
  variances = NULL
)
}
\arguments{
\item{proj_ids}{[vector of integer] Projects IDs from the call.}

\item{total}{[integer] All rows regardless of emptiness.}

\item{columns}{[vector of string] Column names from the call.}

\item{counts}{[vector of integer] Counts of non-empty values, one per column.}

\item{variances}{[vector of number] Variances of values, one per column.}
}
\description{
A list defining a ProjectSetColumnStatsModel
}
