% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectSummaryModel.R
\name{ProjectSummaryModel}
\alias{ProjectSummaryModel}
\title{ProjectSummaryModel}
\usage{
ProjectSummaryModel(projid, title)
}
\arguments{
\item{projid}{[integer, required] Project unique identifier.}

\item{title}{[string, required] Project's title.}
}
\description{
A list defining a ProjectSummaryModel
}
