% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectTaxoStatsModel.R
\name{ProjectTaxoStatsModel}
\alias{ProjectTaxoStatsModel}
\title{ProjectTaxoStatsModel}
\usage{
ProjectTaxoStatsModel(
  projid,
  used_taxa = NULL,
  nb_unclassified,
  nb_validated,
  nb_dubious,
  nb_predicted
)
}
\arguments{
\item{projid}{[integer, required] The project id.}

\item{used_taxa}{[vector of integer] The taxa/category ids used inside the project. An id of -1 means some unclassified objects.}

\item{nb_unclassified}{[integer, required] The number of unclassified objects inside the project.}

\item{nb_validated}{[integer, required] The number of validated objects inside the project.}

\item{nb_dubious}{[integer, required] The number of dubious objects inside the project.}

\item{nb_predicted}{[integer, required] The number of predicted objects inside the project.}
}
\description{
A list defining a ProjectTaxoStatsModel
}
