% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectUserStatsModel.R
\name{ProjectUserStatsModel}
\alias{ProjectUserStatsModel}
\title{ProjectUserStatsModel}
\usage{
ProjectUserStatsModel(projid = NULL, annotators = NULL, activities = NULL)
}
\arguments{
\item{projid}{[integer] The project id.}

\item{annotators}{[vector of ] The users who ever decided on classification or state of objects.}

\item{activities}{[vector of ] More details on annotators' activities.}
}
\description{
A list defining a ProjectUserStatsModel
}
