% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_taxa_update_from_central.R
\name{pull_taxa_update_from_central}
\alias{pull_taxa_update_from_central}
\title{Pull Taxa Update From Central}
\usage{
pull_taxa_update_from_central()
}
\description{
\strong{Returns what changed in EcoTaxoServer managed tree} and update local tree accordingly. i.e. : the number of inserts as nbr_inserts, updates as nbr_updates and errors as errors.
}
