% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_taxa_stats_in_central.R
\name{push_taxa_stats_in_central}
\alias{push_taxa_stats_in_central}
\title{Push Taxa Stats In Central}
\usage{
push_taxa_stats_in_central()
}
\description{
\strong{Push present instance stats}, into EcoTaxoServer.
}
