% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ml_models.R
\name{query_ml_models}
\alias{query_ml_models}
\title{Query Ml Models}
\usage{
query_ml_models()
}
\description{
\strong{Return the list of machine learning models, which can be used for extracting image features.}
}
