% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_taxa.R
\name{query_taxa}
\alias{query_taxa}
\title{Query Taxa}
\usage{
query_taxa(taxon_id)
}
\arguments{
\item{taxon_id}{integer; Internal, the unique numeric id of this taxon.}
}
\description{
Returns \strong{information about the taxon} corresponding to the given id, including its lineage.
}
