% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_taxa_set.R
\name{query_taxa_set}
\alias{query_taxa_set}
\title{Query Taxa Set}
\usage{
query_taxa_set(ids)
}
\arguments{
\item{ids}{string; The separator between numbers is arbitrary non-digit, e.g. ':', '|' or ','.}
}
\description{
Returns \strong{information about several taxa}, including their lineage.
}
