% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_taxa_usage.R
\name{query_taxa_usage}
\alias{query_taxa_usage}
\title{Query Taxa Usage}
\usage{
query_taxa_usage(taxon_id)
}
\arguments{
\item{taxon_id}{integer; Internal, the unique numeric id of this taxon.}
}
\description{
\strong{Where a given taxon is used.} Only validated uses are returned.
}
