% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassif_project_stats.R
\name{reclassif_project_stats}
\alias{reclassif_project_stats}
\title{Reclassif Project Stats}
\usage{
reclassif_project_stats(project_id)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}
}
\description{
Dig into reclassification logs and \strong{return the associations (source → target) for previous reclassifications.}
}
