% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassify_object_set.R
\name{reclassify_object_set}
\alias{reclassify_object_set}
\title{Reclassify Object Set}
\usage{
reclassify_object_set(project_id, forced_id, reason, ProjectFilters)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}

\item{forced_id}{integer; The new classification Id.}

\item{reason}{string; The reason of this new classification.}

\item{ProjectFilters}{Output of ProjectFilters(); How to reduce project data.}
}
\description{
Regardless of present classification or state, \strong{set the new classification for this object set.} If the filter designates "all with given classification", add a TaxonomyChangeLog entry. \strong{Returns the number of affected objects.}
}
