% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_object_set_to_predicted.R
\name{reset_object_set_to_predicted}
\alias{reset_object_set_to_predicted}
\title{Force Object Set To Predicted}
\usage{
reset_object_set_to_predicted(project_id, ProjectFilters)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}

\item{ProjectFilters}{Output of \code{\link[=ProjectFilters]{ProjectFilters()}}; How to reduce project data.}
}
\description{
\strong{Force to Predicted} all objects for the given project with the filters. Return \strong{NULL upon success.}
}
