% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResetPasswordReq.R
\name{ResetPasswordReq}
\alias{ResetPasswordReq}
\title{ResetPasswordReq}
\usage{
ResetPasswordReq(id = NULL, email = NULL, password = NULL)
}
\arguments{
\item{id}{[integer] User unique identifier.}

\item{email}{[string] Email}

\item{password}{[string] Password}
}
\description{
A list defining a ResetPasswordReq
}
