% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restart_job.R
\name{restart_job}
\alias{restart_job}
\title{Restart Job}
\usage{
restart_job(job_id)
}
\arguments{
\item{job_id}{integer; Internal, the unique numeric id of this job.}
}
\description{
\strong{Restart the job related to the given id.} Return \strong{NULL upon success.} 🔒 The job must be in a restartable state, and be accessible to current user.
}
