% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_query.R
\name{sample_query}
\alias{sample_query}
\title{Sample Query}
\usage{
sample_query(sample_id)
}
\arguments{
\item{sample_id}{integer; Internal, the unique numeric id of this sample.}
}
\description{
Returns \strong{information about the sample} corresponding to the given id.
}
