% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_set_get_stats.R
\name{sample_set_get_stats}
\alias{sample_set_get_stats}
\title{Sample Set Get Stats}
\usage{
sample_set_get_stats(sample_ids)
}
\arguments{
\item{sample_ids}{string; String containing the list of one or more sample ids separated by non-num char.}
}
\description{
Returns \strong{classification statistics} for each sample of the given list. One block of stats is returned for each input ID. EXPECT A SLOW RESPONSE : No cache of such information anywhere.
}
