% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleModel.R
\name{SampleModel}
\alias{SampleModel}
\title{SampleModel}
\usage{
SampleModel(
  sampleid,
  projid,
  orig_id,
  latitude = NULL,
  longitude = NULL,
  dataportal_descriptor = NULL,
  free_columns = NULL
)
}
\arguments{
\item{sampleid}{[integer, required] The sample Id.}

\item{projid}{[integer, required] The project Id.}

\item{orig_id}{[string, required] Original sample ID from initial TSV load.}

\item{latitude}{[number] The latitude.}

\item{longitude}{[number] The longitude.}

\item{dataportal_descriptor}{[string] Dataportal descriptor.}

\item{free_columns}{[list] Free columns from sample mapping in project.}
}
\description{
A list defining a SampleModel
}
