% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_search.R
\name{samples_search}
\alias{samples_search}
\title{Samples Search}
\usage{
samples_search(project_ids, id_pattern)
}
\arguments{
\item{project_ids}{string; String containing the list of one or more project id separated by non-num char.}

\item{id_pattern}{string; Sample id textual pattern. Use * or '' for 'any matches'. Match is case-insensitive.}
}
\description{
\strong{Search for samples.}
}
