% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleTaxoStatsModel.R
\name{SampleTaxoStatsModel}
\alias{SampleTaxoStatsModel}
\title{SampleTaxoStatsModel}
\usage{
SampleTaxoStatsModel(
  sample_id = NULL,
  used_taxa = NULL,
  nb_unclassified = NULL,
  nb_validated = NULL,
  nb_dubious = NULL,
  nb_predicted = NULL
)
}
\arguments{
\item{sample_id}{[integer] The sample id.}

\item{used_taxa}{[vector of integer] The taxa/category ids used inside the sample. -1 for unclassified objects.}

\item{nb_unclassified}{[integer] The number of unclassified objects inside the sample.}

\item{nb_validated}{[integer] The number of validated objects inside the sample.}

\item{nb_dubious}{[integer] The number of dubious objects inside the sample.}

\item{nb_predicted}{[integer] The number of predicted objects inside the sample.}
}
\description{
A list defining a SampleTaxoStatsModel
}
