% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_token.R
\name{save_api_token}
\alias{save_api_token}
\title{Save your EcoTaxa token}
\usage{
save_api_token(token, path = api_token_path())
}
\arguments{
\item{token}{the token string, returned by \verb{[login()]}.}

\item{path}{path to the file where the token will be stored; defaults to \verb{[api_token_path()]}.}
}
\value{
The path to the token file.
}
\description{
The token is saved to a file, its path can be changed with option "ecotaxar.token_path"; by default, it is "~/.R/ecotaxar/token".
}
\examples{
token <- login(username="ecotaxa.api.user@gmail.com", password="test!")
# save to the default path
save_api_token(token)
api_token_path()

# save to another path
save_api_token(token, path=)
options(ecotaxa.token_path="~/.my_ecotaxa_token")
api_token_path()
# set the option to change the path
api_token_path()
# NB: write this option in .Rprofile to make it permanent.
}
