% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_collections.R
\name{search_collections}
\alias{search_collections}
\title{Search Collections}
\usage{
search_collections(title, fields = NULL)
}
\arguments{
\item{title}{string; Search by title, use \% for searching with 'any char'.}

\item{fields}{string; Return the default fields (typically used in conjunction with an additional field list). For users list display purpose.}
}
\description{
\strong{Search for collections.} Note: Only manageable collections are returned.
}
