% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_guest.R
\name{search_guest}
\alias{search_guest}
\title{Search Guest}
\usage{
search_guest(by_name = NULL)
}
\arguments{
\item{by_name}{string; Search by name, use \% for searching with 'any char'.}
}
\description{
\strong{Search guests using various criteria}, search is case-insensitive and might contain \% chars.
}
