% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_organizations.R
\name{search_organizations}
\alias{search_organizations}
\title{Search Organizations}
\usage{
search_organizations(name)
}
\arguments{
\item{name}{string; Search by name, use \% for searching with 'any char'.}
}
\description{
\strong{Search for organizations.}
}
