% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_projects.R
\name{search_projects}
\alias{search_projects}
\title{Search Projects}
\usage{
search_projects(
  also_others = NULL,
  not_granted = NULL,
  for_managing = NULL,
  title_filter = NULL,
  instrument_filter = NULL,
  filter_subset = NULL,
  order_field = NULL,
  fields = NULL,
  window_start = NULL,
  window_size = NULL
)
}
\arguments{
\item{also_others}{boolean; NULL}

\item{not_granted}{boolean; Return projects on which the current user has \emph{no permission}, but visible to him/her.}

\item{for_managing}{boolean; Return projects that can be written to (including erased) by the current user.}

\item{title_filter}{string; Use this pattern for matching returned projects names.}

\item{instrument_filter}{string; Only return projects where this instrument was used.}

\item{filter_subset}{boolean; Only return projects having 'subset' in their names.}

\item{order_field}{string; One of \link{'projid', 'title', 'access', 'status', 'objcount', 'pctvalidated', 'pctclassified', 'classifsettings', 'classiffieldlist', 'popoverfieldlist', 'comments', 'rf_models_used', 'cnn_network_id', 'formulae', 'instrument', 'instrument_url', 'highest_right'}}

\item{fields}{string; Return the default fields (typically used in conjunction with an additional field list). For users list display purpose.}

\item{window_start}{integer; Skip \code{window_start} before returning data.}

\item{window_size}{integer; Return only \code{window_size} lines.}
}
\description{
Returns \strong{projects which the current user has explicit permission to access, with search options.} Note that, for performance reasons, in returned ProjectModels, field 'highest_rank' is NOT valued (unlike in simple query). The same information can be found in 'managers', 'annotators' and 'viewers' lists.
}
