% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_taxa.R
\name{search_taxa}
\alias{search_taxa}
\title{Search Taxa}
\usage{
search_taxa(query, project_id = NULL)
}
\arguments{
\item{query}{string; Use this query for matching returned taxa names.}

\item{project_id}{integer; Internal, numeric id of the project.}
}
\description{
\strong{Search for taxa by name.} Queries can be 'small', i.e. of length ﹤3 and even zero-length. 🔓 For a public, unauthenticated call : - zero-length and small queries always return nothing. - otherwise, a full search is done and results are returned in alphabetical order. 🔒 For an authenticated call : - zero-length queries: return the MRU list in full. - small queries: the MRU list is searched, so that taxa in the recent list are returned, if matching. - otherwise, a full search is done. Results are ordered so that taxa in the project list are in first, and are signalled as such in the response.
}
