% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_user.R
\name{search_user}
\alias{search_user}
\title{Search User}
\usage{
search_user(by_name = NULL)
}
\arguments{
\item{by_name}{string; Search by name, use \% for searching with 'any char'.}
}
\description{
\strong{Search users using various criteria}, search is case insensitive and might contain \% chars.
}
