% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_worms_name.R
\name{search_worms_name}
\alias{search_worms_name}
\title{Search Worms Name}
\usage{
search_worms_name(name)
}
\arguments{
\item{name}{string; NULL}
}
\description{
Information about a single taxon in WoRMS reference, including its lineage.
}
