% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_current_user_prefs.R
\name{set_current_user_prefs}
\alias{set_current_user_prefs}
\title{Set Current User Prefs}
\usage{
set_current_user_prefs(project_id, key, value)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}

\item{key}{string; The preference key, as text.}

\item{value}{string; The value to set this preference to, as text.}
}
\description{
\strong{Sets one preference}, for a project and for the currently authenticated user. Available keys are \strong{cwd}, \strong{img_import} and \strong{filters}. The key disappears if set to empty string. \strong{Returns NULL upon success.}
}
