% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_project_predict_settings.R
\name{set_project_predict_settings}
\alias{set_project_predict_settings}
\title{Set Project Predict Settings}
\usage{
set_project_predict_settings(project_id, settings)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}

\item{settings}{string; The new prediction settings.}
}
\description{
\strong{Update the project's prediction settings}, return \strong{NULL upon success.} 🔒 Unlike during full project update above, which needs high permissions, this entry point is accessible to \strong{project annotators}, as it mirrors the prediction privileges.
}
