% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_current_user.R
\name{show_current_user}
\alias{show_current_user}
\title{Show Current User}
\usage{
show_current_user()
}
\description{
Returns \strong{currently authenticated user's} (i.e. you) information, permissions and last used projects.
}
