% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilaritySearchRsp.R
\name{SimilaritySearchRsp}
\alias{SimilaritySearchRsp}
\title{Similarity Search Response}
\usage{
SimilaritySearchRsp(neighbor_ids, sim_scores, message = NULL)
}
\arguments{
\item{neighbor_ids}{[vector of integer, required] The list of similar objects IDs.}

\item{sim_scores}{[vector of number, required] The list of similarity scores, between 0 and 1. The higher the closer, e.g. 1 for the target_id itself.}

\item{message}{[string] A message to the user. If not 'Success' then some condition prevented the computation.}
}
\description{
A list defining a Similarity Search Response
}
