% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleImportReq.R
\name{SimpleImportReq}
\alias{SimpleImportReq}
\title{Simple import request Model}
\usage{
SimpleImportReq(source_path, values, possible_values = NULL)
}
\arguments{
\item{source_path}{[string, required] Source path on server, to zip or plain directory.}

\item{values}{[list, required] :imgdate, imgtime, latitude, longitude, depthmin, depthmax, taxolb, userlb, datelb, status}

\item{possible_values}{[vector of string] Possible Values}
}
\description{
A list defining a Simple import request Model
}
