% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleImportRsp.R
\name{SimpleImportRsp}
\alias{SimpleImportRsp}
\title{SimpleImportRsp}
\usage{
SimpleImportRsp(job_id, errors)
}
\arguments{
\item{job_id}{[integer, required] The job which was created for the run. 0 if called with dry_run option.}

\item{errors}{[vector of string, required] Validation errors, dry_run or not.}
}
\description{
A list defining a SimpleImportRsp
}
