% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetReq.R
\name{SubsetReq}
\alias{SubsetReq}
\title{Subset request Model}
\usage{
SubsetReq(filters = NULL, dest_prj_id, group_type, limit_type, limit_value)
}
\arguments{
\item{filters}{[object of type \link{ProjectFiltersDict}] The filters to apply to project.}

\item{dest_prj_id}{[integer, required] The destination project ID.}

\item{group_type}{[object of type \link{GroupDefinitions}, required] Define the groups in which to apply limits. C for categories, S for samples, A for acquisitions.}

\item{limit_type}{[object of type \link{LimitMethods}, required] The type of limit_value: P for \%, V for constant, both per group.}

\item{limit_value}{[number, required] Limit value, e.g. 20\% or 5 per copepoda or 5\% per sample.}
}
\description{
A list defining a Subset request Model
}
