% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetRsp.R
\name{SubsetRsp}
\alias{SubsetRsp}
\title{SubsetRsp}
\usage{
SubsetRsp(job_id)
}
\arguments{
\item{job_id}{[integer, required] The job created for this operation.}
}
\description{
A list defining a SubsetRsp
}
