% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryExportGroupingEnum.R
\name{SummaryExportGroupingEnum}
\alias{SummaryExportGroupingEnum}
\title{SummaryExportGroupingEnum}
\usage{
SummaryExportGroupingEnum()
}
\description{
A list defining a SummaryExportGroupingEnum
}
