% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryExportQuantitiesOptionsEnum.R
\name{SummaryExportQuantitiesOptionsEnum}
\alias{SummaryExportQuantitiesOptionsEnum}
\title{SummaryExportQuantitiesOptionsEnum}
\format{
string containing abundance, concentration or biovolume
}
\description{
A vector defining a SummaryExportQuantitiesOptionsEnum
}
