% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryExportReq.R
\name{SummaryExportReq}
\alias{SummaryExportReq}
\title{Summary Export request Model}
\usage{
SummaryExportReq(
  project_id,
  quantity = NULL,
  summarise_by = NULL,
  taxo_mapping = NULL,
  formulae = NULL,
  out_to_ftp = NULL
)
}
\arguments{
\item{project_id}{[integer, required] The project to export.}

\item{quantity}{[object of type \link{SummaryExportQuantitiesOptionsEnum}] The quantity to compute. Abundance is always possible.}

\item{summarise_by}{[object of type \link{SummaryExportSumOptionsEnum}] Computations aggregation level.}

\item{taxo_mapping}{[list] Mapping from present taxon (key) to output replacement one (value). Use a 0 replacement to \emph{discard} the present taxon.}

\item{formulae}{[list] Transitory: How to get values from DB free columns. Python syntax, prefixes are 'sam', 'ssm' and 'obj'.Variables used in computations are 'total_water_volume', 'subsample_coef' and 'individual_volume'}

\item{out_to_ftp}{[boolean] Copy result file to FTP area. Original file is still available.}
}
\description{
A list defining a Summary Export request Model
}
