% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryExportSumOptionsEnum.R
\name{SummaryExportSumOptionsEnum}
\alias{SummaryExportSumOptionsEnum}
\title{SummaryExportSumOptionsEnum}
\format{
string containing none, sample or acquisition
}
\description{
A vector defining a SummaryExportSumOptionsEnum
}
