% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_error.R
\name{system_error}
\alias{system_error}
\title{System Error}
\usage{
system_error()
}
\description{
\strong{Return a 500 internal error}, on purpose so the stack trace is visible and client can see what it gives.
}
