% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_status.R
\name{system_status}
\alias{system_status}
\title{System Status}
\usage{
system_status()
}
\description{
\strong{Report the status}, mainly used for verifying that the server is up. Depending on provided credentials, you get more or less information.
}
