% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonCentral.R
\name{TaxonCentral}
\alias{TaxonCentral}
\title{TaxonCentral}
\usage{
TaxonCentral(
  id,
  parent_id = NULL,
  name,
  id_source = NULL,
  taxotype,
  display_name = NULL,
  lastupdate_datetime = NULL,
  id_instance = NULL,
  taxostatus,
  rename_to = NULL,
  source_url = NULL,
  source_desc = NULL,
  creator_email = NULL,
  creation_datetime = NULL,
  nbrobj = NULL,
  nbrobjcum = NULL
)
}
\arguments{
\item{id}{[integer, required] The unique numeric id of the taxon.}

\item{parent_id}{[integer] The unique numeric id of the taxon parent.}

\item{name}{[string, required] The name of the taxon.}

\item{id_source}{[string] The source ID.}

\item{taxotype}{[string, required] The taxon type, 'M' for Morpho or 'P' for Phylo.}

\item{display_name}{[string] The display name of the taxon. It is suffixed in EcoTaxoServer with (Deprecated) when taxostatus is 'D'}

\item{lastupdate_datetime}{[string] Taxon last update. Date, with format YYYY-MM-DD hh:mm:ss.}

\item{id_instance}{[integer] The instance Id.}

\item{taxostatus}{[string, required] The taxon status, N for Not approved, A for Approved or D for Deprecated.}

\item{rename_to}{[integer] The advised replacement Name if the taxon is deprecated.}

\item{source_url}{[string] The source url.}

\item{source_desc}{[string] The source description.}

\item{creator_email}{[string] Email of the creator of the taxon.}

\item{creation_datetime}{[string] Taxon creation date. Date, with format YYYY-MM-DD hh:mm:ss.}

\item{nbrobj}{[integer] Number of objects in this category exactly.}

\item{nbrobjcum}{[integer] Number of objects in this category and descendant ones.}
}
\description{
A list defining a TaxonCentral
}
