% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonomyTreeStatus.R
\name{TaxonomyTreeStatus}
\alias{TaxonomyTreeStatus}
\title{TaxonomyTreeStatus}
\usage{
TaxonomyTreeStatus(last_refresh = NULL)
}
\arguments{
\item{last_refresh}{[string] Taxonomy tree last refresh/sync from taxonomy server. Date, with format YYYY-MM-DDThh:mm:ss.}
}
\description{
A list defining a TaxonomyTreeStatus
}
