% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonUsageModel.R
\name{TaxonUsageModel}
\alias{TaxonUsageModel}
\title{TaxonUsageModel}
\usage{
TaxonUsageModel(projid, title, nb_validated)
}
\arguments{
\item{projid}{[integer, required] Project unique identifier.}

\item{title}{[string, required] Project's title.}

\item{nb_validated}{[integer, required] How many validated objects in this category in this project.}
}
\description{
A list defining a TaxonUsageModel
}
