% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_acquisitions.R
\name{update_acquisitions}
\alias{update_acquisitions}
\title{Update Acquisitions}
\usage{
update_acquisitions(BulkUpdateReq)
}
\arguments{
\item{BulkUpdateReq}{Output of \code{\link[=BulkUpdateReq]{BulkUpdateReq()}};}
}
\description{
Do the required \strong{update for each acquisition in the set}. \strong{Return the number of updated entities.}
}
