% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_collection.R
\name{update_collection}
\alias{update_collection}
\title{Update Collection}
\usage{
update_collection(collection_id, CollectionReq)
}
\arguments{
\item{collection_id}{integer; Internal, the unique numeric id of this collection.}

\item{CollectionReq}{Output of \code{\link[=CollectionReq]{CollectionReq()}}; update full or partial collection Request model}
}
\description{
\strong{Update the collection}. Note that some updates are silently failing when not compatible with the composing projects. \strong{Returns NULL upon success.} Note: The collection is updated only if manageable.
}
