% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_guest.R
\name{update_guest}
\alias{update_guest}
\title{Update Guests}
\usage{
update_guest(guest_id, GuestModel)
}
\arguments{
\item{guest_id}{integer; Internal, numeric id of the guest.}

\item{GuestModel}{Output of \code{\link[=GuestModel]{GuestModel()}};}
}
\description{
\strong{Update the guest}, return \strong{NULL upon success.} 🔒 Depending on logged user, different authorizations apply: - An administrator or user administrator or manager user can change any field with respect of consistency.
}
