% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_organization.R
\name{update_organization}
\alias{update_organization}
\title{Update Organization}
\usage{
update_organization(organization_id, OrganizationModel)
}
\arguments{
\item{organization_id}{integer; Internal, numeric id of the organization.}

\item{OrganizationModel}{Output of \code{\link[=OrganizationModel]{OrganizationModel()}};}
}
\description{
\strong{Update the organization}, return \strong{NULL upon success.} 🔒 Depending on logged user, different authorizations apply: - An administrator or user administrator or manager user can change any field with respect of consistency.
}
