% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_processes.R
\name{update_processes}
\alias{update_processes}
\title{Update Processes}
\usage{
update_processes(BulkUpdateReq)
}
\arguments{
\item{BulkUpdateReq}{Output of \code{\link[=BulkUpdateReq]{BulkUpdateReq()}};}
}
\description{
Do the required \strong{update for each process in the set.} \strong{Returns the number of updated entities.}
}
