% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_project.R
\name{update_project}
\alias{update_project}
\title{Update Project}
\usage{
update_project(project_id, ProjectModel)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}

\item{ProjectModel}{Output of ProjectModel(); Basic and computed information about the Project.}
}
\description{
\strong{Update the project}, return \strong{NULL upon success.} Note that some fields will \strong{NOT} be updated and simply ignored, e.g. \emph{free_cols}.
}
