% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/used_constants.R
\name{used_constants}
\alias{used_constants}
\title{Used Constants}
\usage{
used_constants()
}
\description{
\strong{Return useful strings for user dialog.} Now also used for values extracted from Config.
}
