% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserActivateReq.R
\name{UserActivateReq}
\alias{UserActivateReq}
\title{UserActivateReq}
\usage{
UserActivateReq(token = NULL, reason = NULL, password = NULL)
}
\arguments{
\item{token}{[string] token when the user is not an admin and must confirm the email.}

\item{reason}{[string] status,optional users administrator comment related to the status.}

\item{password}{[string] Existing user can modify own email and must confirm it with token and password when email confirmation is on.}
}
\description{
A list defining a UserActivateReq
}
