% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserActivity.R
\name{UserActivity}
\alias{UserActivity}
\title{UserActivity}
\usage{
UserActivity(id = NULL, nb_actions = NULL, last_annot = NULL)
}
\arguments{
\item{id}{[integer] Id}

\item{nb_actions}{[integer] Nb Actions}

\item{last_annot}{[string] Last Annot}
}
\description{
A list defining a UserActivity
}
