% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidationError.R
\name{ValidationError}
\alias{ValidationError}
\title{ValidationError}
\usage{
ValidationError(loc, msg, type)
}
\arguments{
\item{loc}{[vector of string, required] Location}

\item{msg}{[string, required] Message}

\item{type}{[string, required] Error Type}
}
\description{
A list defining a ValidationError
}
