% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackupExportReq.R
\name{BackupExportReq}
\alias{BackupExportReq}
\title{Backup Export request Model}
\usage{
BackupExportReq(project_id, out_to_ftp = NULL)
}
\arguments{
\item{project_id}{[integer, required] The project to export.}

\item{out_to_ftp}{[boolean] Copy result file to FTP area. Original file is still available.}
}
\description{
A list defining a Backup Export request Model
}
