% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_object_set.R
\name{classify_object_set}
\alias{classify_object_set}
\title{Classify Object Set}
\usage{
classify_object_set(ClassifyReq)
}
\arguments{
\item{ClassifyReq}{Output of ClassifyReq();}
}
\description{
\strong{Change classification and/or qualification for a set of objects.} \strong{Returns the number of updated entities.} 🔒 Current user needs at \emph{least Annotate} right on all projects of specified objects.
}
