% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_collection.R
\name{create_collection}
\alias{create_collection}
\title{Create Collection}
\usage{
create_collection(title = NULL, project_ids = NULL)
}
\arguments{
\item{title}{string; The collection title.}

\item{project_ids}{array; The list of composing project IDs.}
}
\description{
\strong{Create a collection} with at least one project inside. Returns the created collection Id. 🔒 \emph{For admins only.}
}
