% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create Project}
\usage{
create_project(clone_of_id = NULL, title = NULL, visible = NULL)
}
\arguments{
\item{clone_of_id}{integer; Internal, numeric id of a project to clone as a new one. By default it does not clone anything.}

\item{title}{string; The project title, as text.}

\item{visible}{boolean; When TRUE, the project is created visible by all users.}
}
\description{
\strong{Create an empty project with only a title,} and \strong{return the numeric id of this newly created project}. The project will be managed by current user. 🔒 The user has to be \emph{app administrator} or \emph{project creator}.
}
