% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_user.R
\name{create_user}
\alias{create_user}
\title{Create User}
\usage{
create_user(no_bot = NULL, UserModelWithRights)
}
\arguments{
\item{no_bot}{array; not-a-robot proof}

\item{UserModelWithRights}{Output of UserModelWithRights();}
}
\description{
\strong{Create a new user}, return \strong{NULL upon success.} 🔒 Depending on logged user, different authorizations apply: - An administrator or user administrator can create a user. - An unlogged user can self-create an account. But must eventually provide a no-robot proof. - An ordinary logged user cannot create another account. If back-end configuration for self-creation check is Google reCAPTCHA, then no_bot is a pair \link{remote IP, reCAPTCHA response}.
}
