% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateProjectReq.R
\name{CreateProjectReq}
\alias{CreateProjectReq}
\title{Create project request Model}
\usage{
CreateProjectReq(clone_of_id = NULL, title, instrument = NULL, visible = NULL)
}
\arguments{
\item{clone_of_id}{[integer] Internal, numeric id of a project to clone as a new one. By default it does not clone anything.}

\item{title}{[string, required] The project title, as text.}

\item{instrument}{[string] The project instrument.}

\item{visible}{[boolean] When TRUE, the project is created visible by all users.}
}
\description{
A list defining a Create project request Model
}
