% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotaxarapi-package.R
\docType{package}
\name{ecotaxarapi-package}
\alias{ecotaxarapi}
\alias{ecotaxarapi-package}
\title{ecotaxarapi: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
