% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_object_set.R
\name{export_object_set}
\alias{export_object_set}
\title{Export Object Set}
\usage{
export_object_set(filters = NULL, request = NULL)
}
\arguments{
\item{filters}{NULL; NULL}

\item{request}{NULL; NULL}
}
\description{
\strong{Start an export job for the given object set and options.}
}
