% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_object_set.R
\name{export_object_set}
\alias{export_object_set}
\title{Export Object Set}
\usage{
export_object_set(Body_export_object_set_object_set_export_post)
}
\arguments{
\item{Body_export_object_set_object_set_export_post}{Output of \code{\link[=Body_export_object_set_object_set_export_post]{Body_export_object_set_object_set_export_post()}};}
}
\description{
⚠️ Deprecated, see general, summary and backup exports for alternatives. Start an export job for the given object set and options. 🔒 Current user needs \emph{at least Read} right on the requested project.
}
