% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportReq.R
\name{ExportReq}
\alias{ExportReq}
\title{Export request Model}
\usage{
ExportReq(
  collection_id = NULL,
  project_id,
  exp_type,
  use_latin1 = NULL,
  tsv_entities = NULL,
  only_annotations = NULL,
  split_by = NULL,
  coma_as_separator = NULL,
  format_dates_times = NULL,
  with_images = NULL,
  with_internal_ids = NULL,
  with_types_row = NULL,
  only_first_image = NULL,
  sum_subtotal = NULL,
  pre_mapping = NULL,
  formulae = NULL,
  out_to_ftp = NULL
)
}
\arguments{
\item{collection_id}{[integer] The Collection to export if requested.}

\item{project_id}{[integer, string, required] The project(int) or projects (str, project ids list) to export.}

\item{exp_type}{[object of type \link{ExportTypeEnum}, required] The export type.}

\item{use_latin1}{[boolean] Export using latin 1 character set, AKA iso-8859-1. Default is utf-8.}

\item{tsv_entities}{[string] For 'TSV' type, the entities to export, one letter for each of O(bject), P(rocess), A(cquisition), S(ample), C(omments).}

\item{only_annotations}{[boolean] For 'BAK' type, only save objects' last annotation data in backup.}

\item{split_by}{[string] For 'TSV' type, inside archives, split in one directory per... 'sample', 'acquisition', 'taxon' or '' (no split).}

\item{coma_as_separator}{[boolean] For 'TSV' type, use a , instead of . for decimal separator.}

\item{format_dates_times}{[boolean] For 'TSV' type, format dates and times using - and : respectively.}

\item{with_images}{[boolean] For 'BAK' and 'DOI' types, export images as well.}

\item{with_internal_ids}{[boolean] For 'TSV' type, export internal DB IDs.}

\item{with_types_row}{[boolean] Add an EcoTaxa-compatible second line with types.}

\item{only_first_image}{[boolean] For 'DOI' type, export only first (displayed) image.}

\item{sum_subtotal}{[object of type \link{SummaryExportGroupingEnum}] For 'SUM', 'ABO', 'CNC' and 'BIV' types, if computations should be combined. Per A(cquisition) or S(ample) or \if{html}{\out{<Empty>}}(just taxa).}

\item{pre_mapping}{[list] For 'ABO', 'CNC' and 'BIV' types types, mapping from present taxon (key) to output replacement one (value). Use a null replacement to \emph{discard} the present taxon.}

\item{formulae}{[list] Transitory: For 'CNC' and 'BIV' type, how to get values from DB free columns. Python syntax, prefixes are 'sam', 'ssm' and 'obj'.Variables used in computations are 'total_water_volume', 'subsample_coef' and 'individual_volume'}

\item{out_to_ftp}{[boolean] Copy result file to FTP area. Original file is still available.}
}
\description{
A list defining a Export request Model
}
