% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users.R
\name{get_users}
\alias{get_users}
\title{Get Users}
\usage{
get_users(ids = NULL)
}
\arguments{
\item{ids}{string; String containing the list of one or more id separated by non-num char. \strong{If several ids are provided}, one full info is returned per user.}
}
\description{
Returns the list of \strong{all users} with their full information, or just some of them if their ids are provided. 🔒 \emph{For admins only.}
}
