% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_api_response.R
\name{handle_api_response}
\alias{handle_api_response}
\title{Handle an API response}
\usage{
handle_api_response(x)
}
\arguments{
\item{x}{an \link[httr:response]{httr::response} object.}
}
\description{
Fail, with information, if the status code is not 200. Return the parsed response content otherwise.
}
