% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageModel.R
\name{ImageModel}
\alias{ImageModel}
\title{ImageModel}
\usage{
ImageModel(
  imgid = NULL,
  objid,
  imgrank,
  width,
  height,
  orig_file_name,
  thumb_width = NULL,
  thumb_height = NULL,
  file_name,
  thumb_file_name = NULL
)
}
\arguments{
\item{imgid}{[integer] The id of the image.}

\item{objid}{[integer, required] The id of the object related to the image.}

\item{imgrank}{[integer, required] The rank of the image.}

\item{width}{[integer, required] The width of the image.}

\item{height}{[integer, required] The height of the image.}

\item{orig_file_name}{[string, required] The file name of the original image.}

\item{thumb_width}{[integer] Generate thumbnail if image is too large. This generated thumbnail width.}

\item{thumb_height}{[integer] Generate thumbnail if image is too large. The thumb height of the image.}

\item{file_name}{[string, required] The file name.}

\item{thumb_file_name}{[string] If image was too large at import time, the generated thumbnail file name.}
}
\description{
A list defining a ImageModel
}
