% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_file.R
\name{import_file}
\alias{import_file}
\title{Import File}
\usage{
import_file(project_id, ImportReq)
}
\arguments{
\item{project_id}{integer; Internal, numeric id of the project.}

\item{ImportReq}{Output of \code{\link[=ImportReq]{ImportReq()}}; Import request, from UI choices.}
}
\description{
\strong{Validate or do a real import} of an EcoTaxa archive or directory.
}
