% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login}
\usage{
login(password = NULL, username = NULL)
}
\arguments{
\item{password}{string; User password.}

\item{username}{string; User email used during registration.}
}
\description{
\strong{Login barrier,} If successful, the login will returns a \strong{JWT} which will have to be used in bearer authentication scheme for subsequent calls.
}
