% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectModel.R
\name{ObjectModel}
\alias{ObjectModel}
\title{ObjectModel}
\usage{
ObjectModel(
  classif_when = NULL,
  classif_auto_id = NULL,
  classif_auto_score = NULL,
  classif_auto_when = NULL,
  objid,
  acquisid,
  classif_id = NULL,
  objtime = NULL,
  latitude = NULL,
  longitude = NULL,
  depth_min = NULL,
  depth_max = NULL,
  objdate = NULL,
  classif_qual = NULL,
  sunpos = NULL,
  classif_score = NULL,
  classif_who = NULL,
  orig_id,
  object_link = NULL,
  complement_info = NULL,
  sample_id,
  project_id,
  images = NULL,
  free_columns = NULL,
  classif_crossvalidation_id = NULL,
  similarity = NULL,
  random_value
)
}
\arguments{
\item{classif_when}{[string] The human classification date, if \strong{P} or \strong{V}.}

\item{classif_auto_id}{[integer] Set if the object was ever predicted, remains forever with these value. Reflect the 'last state' only if classif_qual is 'P'.}

\item{classif_auto_score}{[number] Set if the object was ever predicted, remains forever with these value. Reflect the 'last state' only if classif_qual is 'P'. The classification auto score is generally between 0 and 1. This is a confidence score, in the fact that, the taxon prediction for this object is correct.}

\item{classif_auto_when}{[string] Set if the object was ever predicted, remains forever with these value. Reflect the 'last state' only if classif_qual is 'P'. The classification date.}

\item{objid}{[integer, required] The object Id.}

\item{acquisid}{[integer, required] The parent acquisition Id.}

\item{classif_id}{[integer] The classification Id.}

\item{objtime}{[string] Object time}

\item{latitude}{[number] The latitude.}

\item{longitude}{[number] The longitude.}

\item{depth_min}{[number] The min depth.}

\item{depth_max}{[number] The min depth.}

\item{objdate}{[string] Object date}

\item{classif_qual}{[string] The classification qualification. Could be \strong{P} for predicted, \strong{V} for validated or \strong{D} for Dubious.}

\item{sunpos}{[string] Sun position, from date, time and coords.}

\item{classif_score}{[number] The ML score for this object, if \strong{P}.}

\item{classif_who}{[integer] The user who manually classified this object, if \strong{V} or \strong{D}.}

\item{orig_id}{[string, required] Original object ID from initial TSV load.}

\item{object_link}{[string] Object link.}

\item{complement_info}{[string] Complement info}

\item{sample_id}{[integer, required] Sample (i.e. parent of parent acquisition) ID.}

\item{project_id}{[integer, required] Project (i.e. parent of sample) ID.}

\item{images}{[vector of ] Images for this object.}

\item{free_columns}{[list] Free columns from object mapping in project.}

\item{classif_crossvalidation_id}{[integer] Always NULL, kept for compat.}

\item{similarity}{[number] Always NULL, kept for compat.}

\item{random_value}{[integer, required] Random value associated to an image}
}
\description{
A list defining a ObjectModel
}
