% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_user_file.R
\name{post_user_file}
\alias{post_user_file}
\title{Put User File}
\usage{
post_user_file()
}
\description{
\strong{Upload a file for the current user.} The returned text will contain a server-side path which is usable for some file-related operations. \emph{e.g. import.}
}
