% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_object_set.R
\name{predict_object_set}
\alias{predict_object_set}
\title{Predict Object Set}
\usage{
predict_object_set(Body_predict_object_set_object_set_predict_post)
}
\arguments{
\item{Body_predict_object_set_object_set_predict_post}{Output of Body_predict_object_set_object_set_predict_post();}
}
\description{
\strong{Start a prediction} AKA automatic classification for the given object set and options.
}
