% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_collections.R
\name{search_collections}
\alias{search_collections}
\title{Search Collections}
\usage{
search_collections(title)
}
\arguments{
\item{title}{string; Search by title, use \% for searching with 'any char'.}
}
\description{
\strong{Search for collections.} 🔒 \emph{For admins only.}
}
