% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryExportQuantitiesOptionsEnum.R
\name{SummaryExportQuantitiesOptionsEnum}
\alias{SummaryExportQuantitiesOptionsEnum}
\title{SummaryExportQuantitiesOptionsEnum}
\usage{
SummaryExportQuantitiesOptionsEnum()
}
\description{
A list defining a SummaryExportQuantitiesOptionsEnum
}
