% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryExportSumOptionsEnum.R
\name{SummaryExportSumOptionsEnum}
\alias{SummaryExportSumOptionsEnum}
\title{SummaryExportSumOptionsEnum}
\usage{
SummaryExportSumOptionsEnum()
}
\description{
A list defining a SummaryExportSumOptionsEnum
}
