% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxaSearchRsp.R
\name{TaxaSearchRsp}
\alias{TaxaSearchRsp}
\title{TaxaSearchRsp}
\usage{
TaxaSearchRsp(id, renm_id = NULL, text, pr)
}
\arguments{
\item{id}{[integer, required] The taxon/category IDs.}

\item{renm_id}{[integer] The advised replacement ID if the taxon/category is deprecated.}

\item{text}{[string, required] The taxon name, display one.}

\item{pr}{[integer, required] 1 if the taxon is in project list, 0 otherwise.}
}
\description{
A list defining a TaxaSearchRsp
}
