% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_acquisitions.R
\name{update_acquisitions}
\alias{update_acquisitions}
\title{Update Acquisitions}
\usage{
update_acquisitions(target_ids = NULL, updates = NULL)
}
\arguments{
\item{target_ids}{array; The IDs of the target entities.}

\item{updates}{array; The list of updates, to do on all impacted entities. { ucol : A column name, pseudo-columns AKA free ones, are OK. uval : The new value to set, always as a string }}
}
\description{
Do the required \strong{update for each acquisition in the set}. \strong{Return the number of updated entities.}
}
