% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_collection.R
\name{update_collection}
\alias{update_collection}
\title{Update Collection}
\usage{
update_collection(
  collection_id,
  project_ids = NULL,
  provider_user = NULL,
  contact_user = NULL,
  creator_users = NULL,
  creator_organisations = NULL,
  associate_users = NULL,
  associate_organisations = NULL,
  id = NULL,
  external_id = NULL,
  external_id_system = NULL,
  title = NULL,
  short_title = NULL,
  citation = NULL,
  license = NULL,
  abstract = NULL,
  description = NULL
)
}
\arguments{
\item{collection_id}{integer; Internal, the unique numeric id of this collection.}

\item{project_ids}{array; The list of composing project IDs.}

\item{provider_user}{NULL; Is the person who is responsible for the content of this metadata record. Writer of the title and abstract.}

\item{contact_user}{NULL; Is the person who should be contacted in cases of questions regarding the content of the dataset or any data restrictions. This is also the person who is most likely to stay involved in the dataset the longest.}

\item{creator_users}{array; All people who are responsible for the creation of the collection. Data creators should receive credit for their work and should therefore be included in the citation.}

\item{creator_organisations}{array; All organisations who are responsible for the creation of the collection. Data creators should receive credit for their work and should therefore be included in the citation.}

\item{associate_users}{array; Other person(s) associated with the collection.}

\item{associate_organisations}{array; Other organisation(s) associated with the collection.}

\item{id}{integer; The collection Id.}

\item{external_id}{string; The external Id.}

\item{external_id_system}{string; The external Id system.}

\item{title}{string; The collection title.}

\item{short_title}{string; The collection short title.}

\item{citation}{string; The collection citation.}

\item{license}{string; The collection license.}

\item{abstract}{string; The collection abstract.}

\item{description}{string; The collection description.}
}
\description{
\strong{Update the collection}. Note that some updates are silently failing when not compatible with the composing projects. \strong{Returns NULL upon success.} Note: The collection is updated only if manageable.
}
