% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_collection_taxonomy_recast.R
\name{update_collection_taxonomy_recast}
\alias{update_collection_taxonomy_recast}
\title{Update Collection Taxo Recast}
\usage{
update_collection_taxonomy_recast(collection_id, TaxonomyRecast)
}
\arguments{
\item{collection_id}{integer; Internal, the unique numeric id of this collection.}

\item{TaxonomyRecast}{Output of TaxonomyRecast(); In various contexts, a taxo recast (from taxon -> to taxon) setting.}
}
\description{
\strong{Create or Update the collection taxonomy recast}. \strong{Returns NULL upon success.} Note: The collection is updated only if manageable.
}
