% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_object_set.R
\name{update_object_set}
\alias{update_object_set}
\title{Update Object Set}
\usage{
update_object_set(BulkUpdateReq)
}
\arguments{
\item{BulkUpdateReq}{Output of \code{\link[=BulkUpdateReq]{BulkUpdateReq()}};}
}
\description{
Do the required \strong{update for each objects in the set.} \strong{Returns the number of updated entities.} 🔒 Current user needs \emph{Manage} right on all projects of specified objects.
}
