% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_samples.R
\name{update_samples}
\alias{update_samples}
\title{Update Samples}
\usage{
update_samples(BulkUpdateReq)
}
\arguments{
\item{BulkUpdateReq}{Output of BulkUpdateReq();}
}
\description{
Do the required \strong{update for each sample in the set.} Any non-null field in the model is written to every impacted sample. \strong{Returns the number of updated entities.}
}
