% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_user.R
\name{update_user}
\alias{update_user}
\title{Update User}
\usage{
update_user(
  user_id,
  id = NULL,
  email = NULL,
  password = NULL,
  name = NULL,
  organisation = NULL,
  active = NULL,
  country = NULL,
  usercreationdate = NULL,
  usercreationreason = NULL,
  can_do = NULL,
  last_used_projects = NULL
)
}
\arguments{
\item{user_id}{integer; Internal, numeric id of the user.}

\item{id}{integer; The unique numeric id of this user.}

\item{email}{string; User's email address, as text, used during registration.}

\item{password}{string; Encrypted (or not) password.}

\item{name}{string; User's full name, as text.}

\item{organisation}{string; User's organisation name, as text.}

\item{active}{boolean; Whether the user is still active.}

\item{country}{string; The country name, as text (but chosen in a consistent list).}

\item{usercreationdate}{string; The date of creation of the user, as text formatted according to the ISO 8601 standard.}

\item{usercreationreason}{string; Paragraph describing the usage of EcoTaxa made by the user.}

\item{can_do}{array; List of User's allowed actions : 1 create a project, 2 administrate the app, 3 administrate users, 4 create taxon.}

\item{last_used_projects}{array; List of User's last used projects.}
}
\description{
\strong{Update the user}, return \strong{NULL upon success.} 🔒 Depending on logged user, different authorizations apply: - An administrator or user administrator can change any field with respect of consistency. - A user can update own password and name. - An ordinary user cannot update anything for another user.
}
